% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{aaa}
\alias{aaa}
\title{Starting H2O For examples}
\description{
Starting H2O For examples
}
\examples{
\dontrun{
if (Sys.info()['sysname'] == "Darwin" && Sys.info()['release'] == '13.4.0') {
  quit(save="no")
} else {
 h2o.init(nthreads = 2)
}
}
}
