% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.character,H2OFrame-method}
\alias{as.character,H2OFrame-method}
\title{Convert H2O Data to Characters}
\usage{
\S4method{as.character}{H2OFrame}(x)
}
\arguments{
\item{x}{a column from an \linkS4class{H2OFrame} data set.
localH2O <- h2o.init()
iris.hex <- as.h2o(iris)
iris.hex[,5] <- as.character(iris.hex[,5])}
}
\description{
Converts an H2O column into character columns.
}

