% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.character.Frame}
\alias{as.character.Frame}
\title{Convert an H2O Frame to a String}
\usage{
\method{as.character}{Frame}(x, ...)
}
\arguments{
\item{x}{An H2O Frame object}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Convert an H2O Frame to a String
}

