% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.data.frame.Frame}
\alias{as.data.frame.Frame}
\title{Converts a Parsed H2O data into a Data Frame}
\usage{
\method{as.data.frame}{Frame}(x, ...)
}
\arguments{
\item{x}{An H2O Frame object.}

\item{...}{Further arguments to be passed down from other methods.}
}
\description{
Downloads the H2O data and then scans it in to an R data frame.
}
\examples{
\donttest{
h2o.init()
prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)
as.data.frame(prostate.hex)
}
}

