% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{as.data.frame.H2OFrame}
\alias{as.data.frame.H2OFrame}
\title{Converts parsed H2O data into an R data frame}
\usage{
\method{as.data.frame}{H2OFrame}(x, ...)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{...}{Further arguments to be passed down from other methods.}
}
\description{
Downloads the H2O data and then scans it in to an R data frame.
}
\details{
Method \code{as.data.frame.H2OFrame} will use \code{\link[data.table]{fread}} if data.table package is installed in required version.
}
\examples{
\dontrun{
library(h2o)
h2o.init()
prostate_path <- system.file("extdata", "prostate.csv", package = "h2o")
prostate <- h2o.uploadFile(path = prostate_path)
as.data.frame(prostate)
}
}
\seealso{
\code{\link{use.package}}
}
