% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.environment,H2OFrame-method}
\alias{as.environment,H2OFrame-method}
\title{Convert H2O Data to an R Enviornment}
\usage{
\S4method{as.environment}{H2OFrame}(x)
}
\arguments{
\item{x}{an \linkS4class{H2OFrame} class object.}
}
\value{
Returns an \code{R} environment object based on the
        \linkS4class{H2OFrame}.
localH2O <- h2o.init()
prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(localH2O, path = prosPath)
names(as.environment)
aa <- as.environment(prostate.hex)
ls(aa)
}
\description{
Converts an \linkS4class{H2OFrame} to an environment.
}

