% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.h2o}
\alias{as.h2o}
\title{R data.frame -> H2OFrame}
\usage{
as.h2o(object, conn = h2o.getConnection(), destination_frame = "")
}
\arguments{
\item{object}{An \code{R} data frame.}

\item{conn}{An \linkS4class{H2OConnection} object containing the IP address and port number
of the H2O server.}

\item{destination_frame}{A string with the desired name for the H2O Frame.}
}
\description{
Import a local R data frame to the H2O cloud.
}

