% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.matrix.Frame}
\alias{as.matrix.Frame}
\title{Convert an H2O Frame to a matrix}
\usage{
\method{as.matrix}{Frame}(x, ...)
}
\arguments{
\item{x}{An H2O Frame object}

\item{...}{Further arguments to be passed down from other methods.}
}
\description{
Convert an H2O Frame to a matrix
}

