% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.numeric,H2OFrame-method}
\alias{as.numeric,H2OFrame-method}
\title{Convert H2O Data to Numeric}
\usage{
\S4method{as.numeric}{H2OFrame}(x)
}
\arguments{
\item{x}{a column from an \linkS4class{H2OFrame} data set.
localH2O <- h2o.init()
prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(localH2O, path = prosPath)
prostate.hex[,2] <- as.factor(prostate.hex[,2])
prostate.hex[,2] <- as.numeric(prostat.hex[,2])}
}
\description{
Converts an H2O column into a numeric value column.
}

