% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.numeric}
\alias{as.numeric}
\title{Convert H2O Data to Numeric}
\usage{
as.numeric(x)
}
\arguments{
\item{x}{a column from an H2O Frame data set.}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Converts an H2O column into a numeric value column.
}
\examples{
\donttest{
h2o.init()
prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)
prostate.hex[,2] <- as.factor (prostate.hex[,2])
prostate.hex[,2] <- as.numeric(prostate.hex[,2])
}
}

