% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.vector}
\alias{as.vector}
\alias{as.vector.Frame}
\title{Convert an H2O Frame to a vector}
\usage{
\\method{as.vector}{Frame}(x,mode)
}
\arguments{
\item{x}{An H2O Frame object}

\item{mode}{Unused}
}
\description{
Convert an H2O Frame to a vector
}

