% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ASTNode-class}
\alias{ASTNode}
\alias{ASTNode-class}
\alias{show,ASTNode-method}
\title{The ASTNode class.}
\usage{
\S4method{show}{ASTNode}(object)
}
\arguments{
\item{object}{An \code{ASTNode} class object.}
}
\description{
This class represents a node in the abstract syntax tree. An ASTNode has a root. The root has children that either
point to another ASTNode, or to a leaf node, which may be of type ASTNumeric or ASTFrame.
}
\section{Slots}{

\describe{
\item{\code{root}}{Object of type \code{Node}}

\item{\code{children}}{Object of type \code{list}}
}}

