% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{australia}
\alias{australia}
\title{Australia Coastal Data}
\format{A data frame with 251 rows and 8 columns}
\description{
Temperature, soil moisture, runoff, and other environmental measurements from
the Australia coast. The data is available from \url{http://cs.colby.edu/courses/S11/cs251/labs/lab07/AustraliaSubset.csv}.
}
\keyword{datasets}

