% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{case_insensitive_match_arg}
\alias{case_insensitive_match_arg}
\title{Works like match.arg but ignores case}
\usage{
case_insensitive_match_arg(arg, choices)
}
\arguments{
\item{arg}{argument to match that should be declared as a character vector containing possible values}

\item{choices}{argument to choose from (OPTIONAL)}
}
\value{
matched arg
}
\description{
Works like match.arg but ignores case
}
