% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dim.Frame}
\alias{dim.Frame}
\title{Returns the Dimensions of an H2O Frame}
\usage{
\method{dim}{Frame}(x)
}
\arguments{
\item{x}{An H2O Frame object.}
}
\description{
Returns the number of rows and columns for a Frame object.
}
\examples{
\donttest{
h2o.init()
iris.hex <- as.h2o(iris)
dim(iris.hex)
}
}
\seealso{
\code{\link[base]{dim}} for the base R method.
}

