% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edicts.R
\name{.addParm}
\alias{.addParm}
\title{TODO: No objects in this file are being used. Either remove file or use objects.}
\usage{
.addParm(parms, k, v)
}
\arguments{
\item{parms}{a list to add the <k,v> pair to}

\item{k}{a key, typically the name of some algorithm parameter}

\item{v}{a value, the value of the algorithm parameter}
}
\description{
Append a <key,value> pair to a list.
}
\details{
Contained here are a set of helper methods that perform type checking on the value passed in.
}
