% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.check_for_ggplot2}
\alias{.check_for_ggplot2}
\title{Stop with a user friendly message if a user is missing the ggplot2 package or has an old version of it.}
\usage{
.check_for_ggplot2(version = "3.0.0")
}
\arguments{
\item{version}{minimal required ggplot2 version}
}
\description{
Stop with a user friendly message if a user is missing the ggplot2 package or has an old version of it.
}
