% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.collapse}
\alias{.collapse}
\title{Helper Collapse Function}
\usage{
.collapse(v)
}
\arguments{
\item{v}{Character vector.}
}
\description{
Collapse a character vector into a ','-sep array of the form: [thing1,thing2,...]
}
