% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.consolidate_varimps}
\alias{.consolidate_varimps}
\title{Consolidate variable importances}
\usage{
.consolidate_varimps(model)
}
\arguments{
\item{model}{H2OModel}
}
\value{
sorted named vector
}
\description{
Consolidation works in the following way:
1. if varimp variable is in x => add it to consolidated_varimps
2. for all remaining varimp variables:
   1. find the longest prefix of varimp variable that is in x and add it to the consolidated varimp
   2. if there was no match, throw an error
3. normalize the consolidated_varimps so they sum up to 1
}
