% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.create_leaderboard}
\alias{.create_leaderboard}
\title{Create a leaderboard like data frame for \code{models}}
\usage{
.create_leaderboard(models_info, leaderboard_frame, top_n = 20)
}
\arguments{
\item{models_info}{H2OAutoML object or list of models}

\item{leaderboard_frame}{when provided with list of models, use this frame to calculate metrics}

\item{top_n}{create leaderboard with just top_n models}
}
\value{
a data.frame
}
\description{
Create a leaderboard like data frame for \code{models}
}
