% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.customized_call}
\alias{.customized_call}
\title{A helper function that makes it easier to override/add params in a function call.}
\usage{
.customized_call(fun, ..., overridable_defaults = NULL, overrides = NULL)
}
\arguments{
\item{fun}{Function to be called}

\item{...}{Parameters that can't be overridden}

\item{overridable_defaults}{List of parameters and values that can be overridden}

\item{overrides}{Parameters to add/override.}
}
\value{
result of \code{fun}
}
\description{
A helper function that makes it easier to override/add params in a function call.
}
