% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.find_appropriate_column_name}
\alias{.find_appropriate_column_name}
\title{Tries to match a \code{fuzzy_col_name} with a column name that exists in \code{cols}.}
\usage{
.find_appropriate_column_name(fuzzy_col_name, cols)
}
\arguments{
\item{fuzzy_col_name}{a name to be decoded}

\item{cols}{vector of columns that contain all possible column names, i.e., decode
fuzzy_col_name must be in cols}
}
\value{
a correct column name
}
\description{
Tries to match a \code{fuzzy_col_name} with a column name that exists in \code{cols}.
}
