% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.get_algorithm}
\alias{.get_algorithm}
\title{Get the algoritm used by the model_or_model_id}
\usage{
.get_algorithm(model_or_model_id, treat_xrt_as_algorithm = FALSE)
}
\arguments{
\item{model_or_model_id}{Model object or a string containing model id}

\item{treat_xrt_as_algorithm}{Try to find out if a model is XRT and if so report it as xrt}
}
\value{
algorithm name
}
\description{
Get the algoritm used by the model_or_model_id
}
