% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.get_domain_mapping}
\alias{.get_domain_mapping}
\title{Get a mapping between columns and their domains}
\usage{
.get_domain_mapping(model)
}
\arguments{
\item{model}{an h2o model}
}
\value{
list containing a mapping from column to its domains (levels)
}
\description{
Get a mapping between columns and their domains
}
