% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.get_feature_count}
\alias{.get_feature_count}
\title{Get feature count sorted by the count descending.}
\usage{
.get_feature_count(column)
}
\arguments{
\item{column}{H2OFrame column}
}
\value{
named vector with feature counts
}
\description{
Get feature count sorted by the count descending.
}
