% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.get_first_of_family}
\alias{.get_first_of_family}
\title{Get first of family models}
\usage{
.get_first_of_family(models, all_stackedensembles = TRUE)
}
\arguments{
\item{models}{models or model ids}

\item{all_stackedensembles}{if TRUE, select all stacked ensembles}
}
\description{
Get first of family models
}
