% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__ALL_CAPABILITIES}
\alias{.h2o.__ALL_CAPABILITIES}
\title{Capabilities endpoints}
\format{
An object of class \code{character} of length 1.
}
\usage{
.h2o.__ALL_CAPABILITIES
}
\description{
Capabilities endpoints
}
\keyword{datasets}
