% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__CREATE_FRAME}
\alias{.h2o.__CREATE_FRAME}
\title{H2OFrame Manipulation}
\format{An object of class \code{character} of length 1.}
\usage{
.h2o.__CREATE_FRAME
}
\description{
H2OFrame Manipulation
}
\keyword{datasets}
