% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{.h2o.__EXPORT_FILES}
\alias{.h2o.__EXPORT_FILES}
\title{Export Files Endpoint Generator}
\usage{
.h2o.__EXPORT_FILES(frame)
}
\arguments{
\item{frame}{H2OFrame}
}
\description{
Export Files Endpoint Generator
}
