% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__FRAMES}
\alias{.h2o.__FRAMES}
\title{Inspect/Summary Endpoints}
\format{
An object of class \code{character} of length 1.
}
\usage{
.h2o.__FRAMES
}
\description{
Inspect/Summary Endpoints
}
\keyword{datasets}
