% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__IMPORT}
\alias{.h2o.__IMPORT}
\title{Import/Export Endpoints}
\format{An object of class \code{character} of length 1.}
\usage{
.h2o.__IMPORT
}
\description{
Import/Export Endpoints
}
\keyword{datasets}
