% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__JOBS}
\alias{.h2o.__JOBS}
\title{Administrative Endpoints}
\format{
An object of class \code{character} of length 1.
}
\usage{
.h2o.__JOBS
}
\description{
Administrative Endpoints
}
\keyword{datasets}
