% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{.h2o.__MODEL_BUILDERS}
\alias{.h2o.__MODEL_BUILDERS}
\title{Model Builder Endpoint Generator}
\usage{
.h2o.__MODEL_BUILDERS(algo)
}
\arguments{
\item{algo}{Cannonical identifier of H2O algorithm.}
}
\description{
Model Builder Endpoint Generator
}
