% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{.h2o.__MODEL_METRICS}
\alias{.h2o.__MODEL_METRICS}
\title{Model Metrics Endpoint}
\usage{
.h2o.__MODEL_METRICS(model, data)
}
\arguments{
\item{model}{H2OModel.}

\item{data}{H2OFrame.}
}
\description{
Model Metrics Endpoint
}
