% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__MODELS}
\alias{.h2o.__MODELS}
\title{Model Endpoint}
\format{
An object of class \code{character} of length 1.
}
\usage{
.h2o.__MODELS
}
\description{
Model Endpoint
}
\keyword{datasets}
