% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__PARSE_SETUP}
\alias{.h2o.__PARSE_SETUP}
\title{Parse Endpoints}
\format{
An object of class \code{character} of length 1.
}
\usage{
.h2o.__PARSE_SETUP
}
\description{
Parse Endpoints
}
\keyword{datasets}
