% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__RAPIDS}
\alias{.h2o.__RAPIDS}
\title{Rapids Endpoint}
\format{An object of class \code{character} of length 1.}
\usage{
.h2o.__RAPIDS
}
\description{
Rapids Endpoint
}
\keyword{datasets}
