% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__REST_API_VERSION}
\alias{.h2o.__REST_API_VERSION}
\title{H2O Package Constants}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.h2o.__REST_API_VERSION
}
\description{
The API endpoints for interacting with H2O via REST are named here.
}
\details{
Additionally, environment variables for the H2O package are named here.
Endpoint Version
}
\keyword{datasets}
