% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{.h2o.__SEGMENT_MODELS_BUILDERS}
\alias{.h2o.__SEGMENT_MODELS_BUILDERS}
\title{Segment Models Builder Endpoint Generator}
\usage{
.h2o.__SEGMENT_MODELS_BUILDERS(algo)
}
\arguments{
\item{algo}{Cannonical identifier of H2O algorithm.}
}
\description{
Segment Models Builder Endpoint Generator
}
