% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.__W2V_SYNONYMS}
\alias{.h2o.__W2V_SYNONYMS}
\title{Word2Vec Endpoints}
\format{
An object of class \code{character} of length 1.
}
\usage{
.h2o.__W2V_SYNONYMS
}
\description{
Word2Vec Endpoints
}
\keyword{datasets}
