% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{.h2o.doGET}
\alias{.h2o.doGET}
\title{Just like doRawGET but fills in the default h2oRestApiVersion if none is provided}
\usage{
.h2o.doGET(h2oRestApiVersion, urlSuffix, parms, ...)
}
\arguments{
\item{h2oRestApiVersion}{(Optional) A version number to prefix to the urlSuffix.  If no version is provided, a default version is chosen for you.}

\item{urlSuffix}{The partial URL suffix to add to the calculated base URL for the instance}

\item{parms}{(Optional) Parameters to include in the request}

\item{...}{(Optional) Additional parameters.}
}
\value{
A list object as described above
}
\description{
Just like doRawGET but fills in the default h2oRestApiVersion if none is provided
}
