% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{.h2o.perfect_auc}
\alias{.h2o.perfect_auc}
\title{Internal function that calculates a precise AUC from given
probabilities and actual responses.}
\usage{
.h2o.perfect_auc(probs, acts)
}
\arguments{
\item{probs}{An \linkS4class{H2OFrame} holding vector of probabilities.}

\item{acts}{An \linkS4class{H2OFrame} holding vector of actuals.}
}
\description{
Note: The underlying implementation is not distributed and can
only handle limited size of data. For internal use only.
}
