% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.h2o.primitives}
\alias{.h2o.primitives}
\title{Map of operations known to H2O}
\format{
An object of class \code{character} of length 39.
}
\usage{
.h2o.primitives
}
\description{
Map of operations known to H2O
}
\keyword{datasets}
