% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.has_model_coefficients}
\alias{.has_model_coefficients}
\title{Has the \code{model} coefficients?}
\usage{
.has_model_coefficients(model)
}
\arguments{
\item{model}{Either a linear model with coefficients => TRUE, or something else => FALSE}
}
\value{
boolean
}
\description{
Has the \code{model} coefficients?
}
