% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.has_varimp}
\alias{.has_varimp}
\title{Has the model variable importance?}
\usage{
.has_varimp(model)
}
\arguments{
\item{model}{model or a string containing model id}
}
\value{
boolean
}
\description{
Has the model variable importance?
}
