% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.interpretable}
\alias{.interpretable}
\title{Is the model considered to be interpretable, i.e., simple enough.}
\usage{
.interpretable(model)
}
\arguments{
\item{model}{model or a string containing model id}
}
\value{
boolean
}
\description{
Is the model considered to be interpretable, i.e., simple enough.
}
