% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.is_h2o_model}
\alias{.is_h2o_model}
\title{Is the \code{model} an H2O model?}
\usage{
.is_h2o_model(model)
}
\arguments{
\item{model}{Either H2O model/model id => TRUE, or something else => FALSE}
}
\value{
boolean
}
\description{
Is the \code{model} an H2O model?
}
