% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.is_h2o_tree_model}
\alias{.is_h2o_tree_model}
\title{Is the \code{model} a Tree-based H2O Model?}
\usage{
.is_h2o_tree_model(model)
}
\arguments{
\item{model}{Either tree-based H2O model/model id => TRUE, or something else => FALSE}
}
\value{
boolean
}
\description{
Is the \code{model} a Tree-based H2O Model?
}
