% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.is_plotting_to_rnotebook}
\alias{.is_plotting_to_rnotebook}
\title{Check if we are plotting in to r notebook.}
\usage{
.is_plotting_to_rnotebook()
}
\value{
boolean
}
\description{
Check if we are plotting in to r notebook.
}
