% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.leaderboard_for_row}
\alias{.leaderboard_for_row}
\title{Enhance leaderboard with per-model predictions.}
\usage{
.leaderboard_for_row(models_info, newdata, row_index, top_n = 20)
}
\arguments{
\item{models_info}{models_info object}

\item{newdata}{H2OFrame}

\item{row_index}{index of the inspected row}

\item{top_n}{leaderboard will contain top_n models}
}
\value{
H2OFrame
}
\description{
Enhance leaderboard with per-model predictions.
}
