% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.model_ids}
\alias{.model_ids}
\title{Get Model Ids}
\usage{
.model_ids(models)
}
\arguments{
\item{models}{list or vector of models/model_ids}
}
\value{
a vector of \code{model_id}s
}
\description{
When provided with list of models it will extract model ids.
When provided with model ids it won't change anything.
Works for mixed list as well.
}
