% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.pkg.env}
\alias{.pkg.env}
\title{The H2O Package Environment}
\format{An object of class \code{environment} of length 4.}
\usage{
.pkg.env
}
\description{
The H2O Package Environment
}
\keyword{datasets}
