% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.plot_varimp}
\alias{.plot_varimp}
\title{Plot variable importances with ggplot2}
\usage{
.plot_varimp(model, top_n = 10)
}
\arguments{
\item{model}{H2OModel}

\item{top_n}{Plot just top_n features}
}
\value{
list of variable importance, groupped variable importance, and variable importance plot
}
\description{
Plot variable importances with ggplot2
}
