% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.shorten_model_ids}
\alias{.shorten_model_ids}
\title{Shortens model ids if possible (iff there will be same amount of unique model_ids as before)}
\usage{
.shorten_model_ids(model_ids)
}
\arguments{
\item{model_ids}{character vector}
}
\value{
character vector
}
\description{
Shortens model ids if possible (iff there will be same amount of unique model_ids as before)
}
