% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{.skip_if_not_developer}
\alias{.skip_if_not_developer}
\title{H2O <-> R Communication and Utility Methods}
\usage{
.skip_if_not_developer()
}
\description{
Collected here are the various methods used by the h2o-R package to communicate with the H2O
backend. There are methods for checking cluster health, polling, and inspecting objects in
the H2O store.
}
