% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.uniformize}
\alias{.uniformize}
\title{Convert to quantiles when provided with numeric vector.
When col is a factor vector assign uniformly value between 0 and 1 to each level.}
\usage{
.uniformize(col)
}
\arguments{
\item{col}{vector}
}
\value{
vector with values between 0 and 1
}
\description{
Convert to quantiles when provided with numeric vector.
When col is a factor vector assign uniformly value between 0 and 1 to each level.
}
