% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{.varimp}
\alias{.varimp}
\title{Get variable importance in a standardized way.}
\usage{
.varimp(model)
}
\arguments{
\item{model}{H2OModel}
}
\value{
A named vector
}
\description{
Get variable importance in a standardized way.
}
