% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{.verify_dataxy}
\alias{.verify_dataxy}
\title{Used to verify data, x, y and turn into the appropriate things}
\usage{
.verify_dataxy(data, x, y, autoencoder = FALSE)
}
\arguments{
\item{data}{H2OFrame}

\item{x}{features}

\item{y}{response}

\item{autoencoder}{autoencoder flag}
}
\description{
Used to verify data, x, y and turn into the appropriate things
}
