% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{feature_frequencies.H2OModel}
\alias{feature_frequencies.H2OModel}
\alias{h2o.feature_frequencies}
\title{Retrieve the number of occurrences of each feature for given observations 
Available for GBM, Random Forest and Isolation Forest models.}
\usage{
feature_frequencies.H2OModel(object, newdata, ...)

h2o.feature_frequencies(object, newdata, ...)
}
\arguments{
\item{object}{a fitted \linkS4class{H2OModel} object for which prediction is
desired}

\item{newdata}{An H2OFrame object in which to look for
variables with which to predict.}

\item{...}{additional arguments to pass on.}
}
\value{
Returns an H2OFrame contain per-feature frequencies on the predict path for each input row.
}
\description{
Retrieve the number of occurrences of each feature for given observations 
Available for GBM, Random Forest and Isolation Forest models.
}
\seealso{
\code{\link{h2o.gbm}} and  \code{\link{h2o.randomForest}} for model
         generation in h2o.
}
