% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Frame-Extract}
\alias{$.Frame}
\alias{$<-.Frame}
\alias{Frame-Extract}
\alias{[,Frame-method}
\alias{[.Frame}
\alias{[<-.Frame}
\alias{[[.Frame}
\alias{[[<-.Frame}
\title{Extract or Replace Parts of an H2O Frame Object}
\usage{
\method{[}{Frame}(data, row, col, drop = TRUE)

\method{$}{Frame}(x, name)

\method{[[}{Frame}(x, i, exact = TRUE)

\method{$}{Frame}(x, name)

\method{[[}{Frame}(x, i, exact = TRUE)

\method{[}{Frame}(data, row, col, ...) <- value

\method{$}{Frame}(data, name) <- value

\method{[[}{Frame}(data, name) <- value
}
\arguments{
\item{data}{object from which to extract element(s) or in which to replace element(s).}

\item{row}{index specifying row element(s) to extract or replace. Indices are numeric or
character vectors or empty (missing) or will be matched to the names.}

\item{col}{index specifying column element(s) to extract or replace.}

\item{drop}{Unused}

\item{x}{An H2O Frame}

\item{name}{a literal character string or a name (possibly backtick quoted).}

\item{i}{index}

\item{exact}{controls possible partial matching of \code{[[} when extracting
a character}

\item{...}{Further arguments passed to or from other methods.}

\item{value}{To be assigned}
}
\description{
Operators to extract or replace parts of Frame objects.
}

