% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Ops.Frame}
\alias{!.Frame}
\alias{Math.Frame}
\alias{Ops.Frame}
\alias{Summary.Frame}
\alias{\%*\%}
\alias{colnames<-}
\alias{h2o.length}
\alias{is.na.Frame}
\alias{length.Frame}
\alias{log}
\alias{names<-.Frame}
\alias{ncol.Frame}
\alias{nrow.Frame}
\alias{t.Frame}
\alias{trunc}
\title{S3 Group Generic Functions for H2O}
\usage{
\method{Ops}{Frame}(e1, e2)

\method{Math}{Frame}(x, ...)

\method{Math}{Frame}(x, ...)

\method{Math}{Frame}(x, ...)

\method{Summary}{Frame}(x, ..., na.rm)

\method{!}{Frame}(x)

\method{is.na}{Frame}(x)

\method{t}{Frame}(x)

log(x, ...)

trunc(x, ...)

x \%*\% y

nrow.Frame(x)

ncol.Frame(x)

\method{length}{Frame}(x)

h2o.length(x)

\method{names}{Frame}(x) <- value

colnames(x) <- value
}
\arguments{
\item{e1}{object}

\item{e2}{object}

\item{x}{object}

\item{...}{Further arguments passed to or from other methods.}

\item{na.rm}{logical. whether or not missing values should be removed}

\item{y}{object}

\item{value}{To be assigned}
}
\description{
Methods for group generic functions and H2O objects.
}

