% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{generate_col_ind}
\alias{generate_col_ind}
\title{CHeck to see if the column names/indices entered is valid for the dataframe given.  This is an internal function}
\usage{
generate_col_ind(data, by)
}
\arguments{
\item{data}{The H2OFrame whose column names or indices are entered as a list}

\item{by}{The column names/indices in a list.}
}
\description{
CHeck to see if the column names/indices entered is valid for the dataframe given.  This is an internal function
}
