% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.abs}
\alias{h2o.abs}
\title{Compute the absolute value of x}
\usage{
h2o.abs(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the absolute value of x
}
\seealso{
\code{\link[base]{abs}} for the base R implementation.
}

