% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.acos}
\alias{h2o.acos}
\title{Compute the arc cosine of x}
\usage{
h2o.acos(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the arc cosine of x
}
\seealso{
\code{\link[base]{acos}} for the base R implementation.
}

