% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregator.R
\name{h2o.aggregated_frame}
\alias{h2o.aggregated_frame}
\title{Retrieve an aggregated frame from the Aggregator model}
\usage{
h2o.aggregated_frame(model)
}
\arguments{
\item{model}{an \linkS4class{H2OClusteringModel} corresponding from a \code{h2o.aggregator} call.}
}
\description{
Retrieve an aggregated frame from the Aggregator model
}

