% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.all}
\alias{h2o.all}
\title{Given a set of logical vectors, are all of the values true?}
\usage{
h2o.all(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Given a set of logical vectors, are all of the values true?
}
\seealso{
\code{\link[base]{all}} for the base R implementation.
}

