% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.arrange}
\alias{h2o.arrange}
\title{Sorts H2OFrame by the columns specified. Returns a new H2OFrame, like dplyr::arrange.}
\usage{
h2o.arrange(x, ...)
}
\arguments{
\item{x}{The H2OFrame input to be sorted.}

\item{\dots}{The column names to sort by.}
}
\description{
Sorts H2OFrame by the columns specified. Returns a new H2OFrame, like dplyr::arrange.
}

