% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.ascharacter}
\alias{h2o.ascharacter}
\title{Convert H2O Data to Characters}
\usage{
h2o.ascharacter(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Convert H2O Data to Characters
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.ascharacter(iris["species"])
}
}
