% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.asfactor}
\alias{h2o.asfactor}
\title{Convert H2O Data to Factors}
\usage{
h2o.asfactor(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Convert H2O Data to Factors
}
\seealso{
\code{\link[base]{as.factor}} for the base R implementation.
}

