% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.asnumeric}
\alias{h2o.asnumeric}
\title{Convert H2O Data to Numerics}
\usage{
h2o.asnumeric(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
If the column type is enum and you want to convert it to numeric, you should first convert it to character then convert it to numeric. 
Otherwise, the values may be converted to underlying factor values, not the expected mapped values.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/junit/cars_20mpg.csv"
cars <- h2o.importFile(f)
h2o.ascharacter(cars)
h2o.asnumeric(cars)
}
}
\seealso{
\code{\link[base]{numeric}} for the base R implementation, \code{as.numeric()}.
}
