% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.assign}
\alias{h2o.assign}
\title{Rename an H2O object.}
\usage{
h2o.assign(data, key)
}
\arguments{
\item{data}{An H2OFrame object}

\item{key}{The hex key to be associated with the H2O parsed data object}
}
\description{
Makes a copy of the data frame and gives it the desired the key.
}

