% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.biases}
\alias{h2o.biases}
\title{Return the respective bias vector}
\usage{
h2o.biases(object, vector_id = 1)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}}

\item{vector_id}{An integer, ranging from 1 to number of layers + 1, that specifies the bias vector to return.}
}
\description{
Return the respective bias vector
}

