% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.bottomN}
\alias{h2o.bottomN}
\title{H2O bottomN}
\usage{
h2o.bottomN(x, column, nPercent)
}
\arguments{
\item{x}{an H2OFrame}

\item{column}{is a column name or column index to grab the top N percent value from}

\item{nPercent}{is a bottom percentage value to grab}
}
\value{
An H2OFrame with 2 columns.  The first column is the original row indices, second column contains the bottomN values
}
\description{
bottomN function will will grab the bottom N percent of values of a column and return it in a H2OFrame.
Extract the top N percent of values of a column and return it in a H2OFrame.
}

