% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.ceiling}
\alias{h2o.ceiling}
\title{ceiling takes a single numeric argument x and returns a
numeric vector containing the smallest integers not less than the
corresponding elements of x.}
\usage{
h2o.ceiling(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
ceiling takes a single numeric argument x and returns a
numeric vector containing the smallest integers not less than the
corresponding elements of x.
}
\seealso{
\code{\link[base]{ceiling}} for the base R implementation.
}

