% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.centers}
\alias{h2o.centers}
\title{Retrieve the Model Centers}
\usage{
h2o.centers(object, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}

\item{\dots}{further arguments to be passed on (currently unimplemented)}
}
\description{
Retrieve the Model Centers
}

