% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.centroid_stats}
\alias{h2o.centroid_stats}
\title{Retrieve the centroid statistics}
\usage{
h2o.centroid_stats(object, valid = FALSE, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}

\item{valid}{Retrieve the validation metric.}

\item{\dots}{further arguments to be passed on (currently unimplemented)}
}
\description{
Retrieve the centroid statistics
}

