% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.clearLog}
\alias{h2o.clearLog}
\title{Delete All H2O R Logs}
\usage{
h2o.clearLog()
}
\description{
Clear all H2O R command and error response logs from the local disk. Used
primarily for debugging purposes.
}
\examples{
\dontrun{
library(h2o)
localH2O = h2o.init()
h2o.startLogging()
ausPath = system.file("extdata", "australia.csv", package="h2o")
australia.hex = h2o.importFile(localH2O, path = ausPath)
h2o.stopLogging()
h2o.clearLog()
}
}
\seealso{
\code{\link{h2o.startLogging}, \link{h2o.stopLogging},
         \link{h2o.openLog}}
}

