% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.cluster_sizes}
\alias{h2o.cluster_sizes}
\title{Retrieve the cluster sizes}
\usage{
h2o.cluster_sizes(object, valid = FALSE, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}

\item{valid}{Retrieve the validation metric.}

\item{\dots}{further arguments to be passed on (currently unimplemented)}
}
\description{
Retrieve the cluster sizes
}

