% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.clusterIsUp}
\alias{h2o.clusterIsUp}
\title{Determine if an H2O cluster is up or not}
\usage{
h2o.clusterIsUp(conn = h2o.getConnection())
}
\arguments{
\item{conn}{H2O connection object}
}
\value{
TRUE if the cluster is up; FALSE otherwise
}
\description{
Determine if an H2O cluster is up or not
}

