% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.clusterStatus}
\alias{h2o.clusterStatus}
\title{Return the status of the cluster}
\usage{
h2o.clusterStatus()
}
\description{
Retrieve information on the status of the cluster running H2O.
}
\examples{
\donttest{
h2o.init()
h2o.clusterStatus()
}
}
\seealso{
\linkS4class{H2OConnection}, \code{\link{h2o.init}}
}

