% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.coef_norm}
\alias{h2o.coef_norm}
\title{Return coefficients fitted on the standardized data (requires standardize = True, which is on by default). These coefficients can be used to evaluate variable importance.}
\usage{
h2o.coef_norm(object)
}
\arguments{
\item{object}{an \linkS4class{H2OModel} object.}
}
\description{
Return coefficients fitted on the standardized data (requires standardize = True, which is on by default). These coefficients can be used to evaluate variable importance.
}

