% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.colnames}
\alias{h2o.colnames}
\title{Return column names of an H2OFrame}
\usage{
h2o.colnames(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Return column names of an H2OFrame
}
\examples{
\dontrun{
library(h2o)
h2o.init()

frame <- h2o.createFrame(rows = 6, cols = 2,
                         categorical_fraction = 0.0, 
                         missing_fraction = 0.7, 
                         seed = 123)
h2o.colnames(frame)
}
}
\seealso{
\code{\link[base]{colnames}} for the base R implementation.
}
