% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.computeGram}
\alias{h2o.computeGram}
\title{Compute weighted gram matrix.}
\usage{
h2o.computeGram(X, weights = "", use_all_factor_levels = FALSE,
  standardize = TRUE, skip_missing = FALSE)
}
\arguments{
\item{X}{an \linkS4class{H2OModel} corresponding to H2O framel.}

\item{weights}{character corresponding to name of weight vector in frame.}

\item{use_all_factor_levels}{logical flag telling h2o whether or not to skip first level of categorical variables during one-hot encoding.}

\item{standardize}{logical flag telling h2o whether or not to standardize data}

\item{skip_missing}{logical flag telling h2o whether skip rows with missing data or impute them with mean}
}
\description{
Compute weighted gram matrix.
}

