% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.cos}
\alias{h2o.cos}
\title{Compute the cosine of x}
\usage{
h2o.cos(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the cosine of x
}
\seealso{
\code{\link[base]{cos}} for the base R implementation.
}
