% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.cross_validation_holdout_predictions}
\alias{h2o.cross_validation_holdout_predictions}
\title{Retrieve the cross-validation holdout predictions}
\usage{
h2o.cross_validation_holdout_predictions(object)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}
}
\value{
Returns a H2OFrame
}
\description{
Retrieve the cross-validation holdout predictions
}

