% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.cross_validation_predictions}
\alias{h2o.cross_validation_predictions}
\title{Retrieve the cross-validation predictions}
\usage{
h2o.cross_validation_predictions(object)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}
}
\value{
Returns a list of H2OFrame objects
}
\description{
Retrieve the cross-validation predictions
}
