% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.cummax}
\alias{h2o.cummax}
\title{Cumulative Max}
\usage{
h2o.cummax(x)
}
\arguments{
\item{x}{An \linkS4class{H2OFrame} object.}
}
\description{
Obtain the cumulative max of a column.
}
\examples{
localH2O <- h2o.init()
fr <- as.h2o(iris)
h2o.cummax(fr[,1])
}

