% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.cummin}
\alias{h2o.cummin}
\title{Cumulative Min}
\usage{
h2o.cummin(x)
}
\arguments{
\item{x}{An \linkS4class{H2OFrame} object.}
}
\description{
Obtain the cumulative min of a column.
}
\examples{
localH2O <- h2o.init()
fr <- as.h2o(iris)
h2o.cummin(fr[,1])
}

