% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.cumsum}
\alias{h2o.cumsum}
\title{Return the cumulative sum over a column or across a row}
\usage{
h2o.cumsum(x, axis = 0)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{axis}{An int that indicates whether to do down a column (0) or across a row (1).}
}
\description{
Return the cumulative sum over a column or across a row
}
\seealso{
\code{\link[base]{cumsum}} for the base R implementation.
}
