% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.day}
\alias{day}
\alias{day.Frame}
\alias{h2o.day}
\title{Convert Milliseconds to Day of Month in H2O Datasets}
\usage{
h2o.day(x)

day(x)

\method{day}{Frame}(x)
}
\arguments{
\item{x}{An H2O Frame object.}
}
\value{
A Frame object containing the entries of \code{x} converted to days of
        the month.
}
\description{
Converts the entries of a Frame object from milliseconds to days of the month
(on a 1 to 31 scale).
}
\seealso{
\code{\link{h2o.month}}
}

