% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.dimnames}
\alias{h2o.dimnames}
\title{Column names of an H2OFrame}
\usage{
h2o.dimnames(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Column names of an H2OFrame
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/junit/cars_20mpg.csv"
cars <- h2o.importFile(f)
h2o.dimnames(cars)
}
}
\seealso{
\code{\link[base]{dimnames}} for the base R implementation.
}
