% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.downloadAllLogs}
\alias{h2o.downloadAllLogs}
\title{Download H2O Log Files to Disk}
\usage{
h2o.downloadAllLogs(dirname = ".", filename = NULL)
}
\arguments{
\item{dirname}{(Optional) A character string indicating the directory that the log file should be saved in.}

\item{filename}{(Optional) A character string indicating the name that the log file should be saved to.}
}
\description{
\code{h2o.downloadAllLogs} downloads all H2O log files to local disk. Generally used for debugging purposes.
}

