% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.entropy}
\alias{h2o.entropy}
\title{Shannon entropy}
\usage{
h2o.entropy(x)
}
\arguments{
\item{x}{The column on which to calculate the entropy.}
}
\description{
Return the Shannon entropy of a string column. If the string is empty, the entropy is 0.
}
\examples{
\donttest{
library(h2o)
h2o.init()
buys <- as.h2o(c("no", "no", "yes", "yes", "yes", "no", "yes", "no", "yes", "yes","no"))
buys_entropy <- h2o.entropy(buys)
}
}

