% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.exportHDFS}
\alias{h2o.exportHDFS}
\title{Export a H2OFrame to HDFS}
\usage{
h2o.exportHDFS(data, path, force = FALSE)
}
\arguments{
\item{data}{an \linkS4class{H2OFrame} class object.}

\item{path}{The path to write the data to. Must include the driectory and
filename.}

\item{force}{logical, indicates how to deal with files that already exist.}
}
\description{
Exports an \linkS4class{H2OFrame} to HDFS.
}

