% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.find_threshold_by_max_metric}
\alias{h2o.find_threshold_by_max_metric}
\title{Find the threshold, give the max metric}
\usage{
h2o.find_threshold_by_max_metric(object, metric)
}
\arguments{
\item{object}{H2OBinomialMetrics}

\item{metric}{"F1," for example}
}
\description{
Find the threshold, give the max metric
}

