% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.gains_lift_plot,H2OModel-method}
\alias{h2o.gains_lift_plot,H2OModel-method}
\title{Plot Gains/Lift curves}
\usage{
\S4method{h2o.gains_lift_plot}{H2OModel}(object, type = c("both", "gains", "lift"), xval = FALSE)
}
\arguments{
\item{object}{H2OModel object}

\item{type}{What curve to plot. One of "both", "gains", "lift".}

\item{xval}{if TRUE, use cross-validation metrics}
}
\description{
Plot Gains/Lift curves
}
