% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.gains_lift_plot,H2OModelMetrics-method}
\alias{h2o.gains_lift_plot,H2OModelMetrics-method}
\title{Plot Gains/Lift curves}
\usage{
\S4method{h2o.gains_lift_plot}{H2OModelMetrics}(object, type = c("both", "gains", "lift"))
}
\arguments{
\item{object}{H2OModelMetrics object}

\item{type}{What curve to plot. One of "both", "gains", "lift".}
}
\description{
Plot Gains/Lift curves
}
