% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection.R
\name{h2o.get_best_model_predictors}
\alias{h2o.get_best_model_predictors}
\title{Extracts the subset of predictor names that yield the best R2 value for each predictor subset size.}
\usage{
h2o.get_best_model_predictors(model)
}
\arguments{
\item{model}{is a H2OModel with algorithm name of modelselection}
}
\description{
Extracts the subset of predictor names that yield the best R2 value for each predictor subset size.
}
