% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection.R
\name{h2o.get_best_r2_values}
\alias{h2o.get_best_r2_values}
\title{Extracts the best R2 values for all predictor subset size.}
\usage{
h2o.get_best_r2_values(model)
}
\arguments{
\item{model}{is a H2OModel with algorithm name of modelselection}
}
\description{
Extracts the best R2 values for all predictor subset size.
}
