% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\name{h2o.get_gam_knot_column_names}
\alias{h2o.get_gam_knot_column_names}
\title{Extracts the gam column names corresponding to the knot locations from model output if it is enabled.}
\usage{
h2o.get_gam_knot_column_names(model)
}
\arguments{
\item{model}{is a H2OModel with algorithm name of gam}
}
\description{
Extracts the gam column names corresponding to the knot locations from model output if it is enabled.
}
