% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam.R
\name{h2o.get_knot_locations}
\alias{h2o.get_knot_locations}
\title{Extracts the knot locations from model output if it is enabled.}
\usage{
h2o.get_knot_locations(model, gam_column = NULL)
}
\arguments{
\item{model}{is a H2OModel with algorithm name of gam}

\item{gam_column}{will only extract the knot locations for the specific gam_columns.  Else, return all.}
}
\description{
Extracts the knot locations from model output if it is enabled.
}
