% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.get_ntrees_actual}
\alias{h2o.get_ntrees_actual}
\title{Retrieve actual number of trees for tree algorithms}
\usage{
h2o.get_ntrees_actual(object)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}
}
\description{
Retrieve actual number of trees for tree algorithms
}
