% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection.R
\name{h2o.get_predictors_removed_per_step}
\alias{h2o.get_predictors_removed_per_step}
\title{Extracts the predictor removed to model at each step.}
\usage{
h2o.get_predictors_removed_per_step(model)
}
\arguments{
\item{model}{is a H2OModel with algorithm name of modelselection}
}
\description{
Extracts the predictor removed to model at each step.
}
