% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{h2o.getConnection}
\alias{h2o.getConnection}
\title{Retrieve an H2O Connection}
\usage{
h2o.getConnection()
}
\value{
Returns an \linkS4class{H2OConnection} object.
}
\description{
Attempt to recover an h2o connection.
}
