% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.getFrame}
\alias{h2o.getFrame}
\title{Get an R Reference to an H2O Dataset}
\usage{
h2o.getFrame(frame_id, conn = h2o.getConnection(), linkToGC = FALSE)
}
\arguments{
\item{frame_id}{A string indicating the unique frame of the dataset to retrieve.}

\item{conn}{\linkS4class{H2OConnection} object containing the IP address and port
of the server running H2O.}

\item{linkToGC}{a logical value indicating whether to remove the underlying frame
from the H2O cluster when the R proxy object is garbage collected.}
}
\description{
Get the reference to a frame with the given frame_id in the H2O instance.
}

