% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.getFutureModel}
\alias{h2o.getFutureModel}
\title{Get future model}
\usage{
h2o.getFutureModel(object, verbose = FALSE)
}
\arguments{
\item{object}{H2OModel}

\item{verbose}{Print model progress to console. Default is FALSE}
}
\description{
Get future model
}

