% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.getGLMModel}
\alias{h2o.getGLMModel}
\title{Resolve a GLM H2O Futures Model}
\usage{
h2o.getGLMModel(keys, conn)
}
\arguments{
\item{keys}{an \linkS4class{H2OModelFuture} or correct job key.}

\item{conn}{a corresponding \linkS4class{H2OConnection} class object.}
}
\value{
Returns the correct \linkS4class{H2OModel} for the created model.
}
\description{
Turns an \linkS4class{H2OModelFuture} into a model of the correct type.
}

