% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.getId}
\alias{h2o.getId}
\title{Get back-end distributed key/value store id from a Frame.}
\usage{
h2o.getId(x)
}
\arguments{
\item{x}{A Frame}
}
\value{
The id
}
\description{
Get back-end distributed key/value store id from a Frame.
}

