% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{h2o.getLambdaBest}
\alias{h2o.getLambdaBest}
\title{Extract best lambda value found from glm model.}
\usage{
h2o.getLambdaBest(model)
}
\arguments{
\item{model}{an \linkS4class{H2OModel} corresponding from a \code{h2o.glm} call.}
}
\description{
This function allows setting betas of an existing glm model.
}
