% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{h2o.getLambdaMin}
\alias{h2o.getLambdaMin}
\title{Extract the minimum lambda value calculated during lambda search from glm model.
Note that due to early stop, this minimum lambda value may not be used in the actual lambda search.}
\usage{
h2o.getLambdaMin(model)
}
\arguments{
\item{model}{an \linkS4class{H2OModel} corresponding from a \code{h2o.glm} call.}
}
\description{
This function allows setting betas of an existing glm model.
}
