% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.getTimezone}
\alias{h2o.getTimezone}
\title{Get the Time Zone on the H2O Cloud}
\usage{
h2o.getTimezone(conn = h2o.getConnection())
}
\arguments{
\item{conn}{An H2OConnection object.}
}
\description{
Get the Time Zone on the H2O Cloud
}

