% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.gsub}
\alias{h2o.gsub}
\title{String Global Substitute}
\usage{
h2o.gsub(pattern, replacement, x, ignore.case = FALSE)
}
\arguments{
\item{pattern}{The pattern to replace.}

\item{replacement}{The replacement pattern.}

\item{x}{The column on which to operate.}

\item{ignore.case}{Case sensitive or not}
}
\description{
Creates a copy of the target column in which each string has all occurence of
the regex pattern replaced with the replacement substring.
}
\examples{
\donttest{
library(h2o)
h2o.init()
string_to_gsub <- as.h2o("r tutorial")
sub_string <- h2o.gsub("r ","H2O ",string_to_gsub)
}
}

