% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.HGLMMetrics}
\alias{h2o.HGLMMetrics}
\title{Retrieve HGLM ModelMetrics}
\usage{
h2o.HGLMMetrics(object)
}
\arguments{
\item{object}{an H2OModel object or H2OModelMetrics.}
}
\description{
Retrieve HGLM ModelMetrics
}
