% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.hit_ratio_table}
\alias{h2o.hit_ratio_table}
\title{Retrieve the Hit Ratios}
\usage{
h2o.hit_ratio_table(object, valid = FALSE, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}

\item{valid}{Retreive the validation metric.}

\item{\dots}{further arguments to be passed on (currently unimplemented)}
}
\description{
Retrieve the Hit Ratios
}

