% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.hour}
\alias{h2o.hour}
\alias{hour}
\alias{hour.Frame}
\title{Convert Milliseconds to Hour of Day in H2O Datasets}
\usage{
h2o.hour(x)

hour(x)

\method{hour}{Frame}(x)
}
\arguments{
\item{x}{An H2O Frame object.}
}
\value{
A Frame object containing the entries of \code{x} converted to hours of
        the day.
}
\description{
Converts the entries of a Frame object from milliseconds to hours of the day
(on a 0 to 23 scale).
}
\seealso{
\code{\link{h2o.day}}
}

