% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.ischaracter}
\alias{h2o.ischaracter}
\title{Check if character}
\usage{
h2o.ischaracter(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Check if character
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/iris/iris_wheader.csv"
iris <- h2o.importFile(f)
iris_char <- h2o.ascharacter(iris["class"])
h2o.ischaracter(iris_char)
}
}
\seealso{
\code{\link[base]{is.character}} for the base R implementation.
}
