% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.isfactor}
\alias{h2o.isfactor}
\title{Check if factor}
\usage{
h2o.isfactor(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Check if factor
}
\seealso{
\code{\link[base]{is.factor}} for the base R implementation.
}
