% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.isnumeric}
\alias{h2o.isnumeric}
\title{Check if numeric}
\usage{
h2o.isnumeric(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Check if numeric
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/iris/iris_wheader.csv"
iris <- h2o.importFile(f)
h2o.isnumeric(iris["sepal_len"])
}
}
\seealso{
\code{\link[base]{is.numeric}} for the base R implementation.
}
