% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{h2o.keyof}
\alias{h2o.keyof}
\alias{h2o.keyof,Keyed-method}
\alias{h2o.keyof,H2OModel-method}
\alias{h2o.keyof,H2OFrame-method}
\alias{h2o.keyof,H2OAutoML-method}
\title{Method on \code{Keyed} objects allowing to obtain their key.}
\usage{
h2o.keyof(object)

\S4method{h2o.keyof}{Keyed}(object)

\S4method{h2o.keyof}{H2OModel}(object)

\S4method{h2o.keyof}{H2OFrame}(object)

\S4method{h2o.keyof}{H2OAutoML}(object)
}
\arguments{
\item{object}{A \code{Keyed} object}
}
\value{
the string key holding the persistent object.
}
\description{
Method on \code{Keyed} objects allowing to obtain their key.
}
