% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.killMinus3}
\alias{h2o.killMinus3}
\title{Dump the stack into the JVM's stdout.}
\usage{
h2o.killMinus3(conn = h2o.getConnection())
}
\arguments{
\item{conn}{an \linkS4class{H2OConnection} class object.}
}
\description{
A poor man's profiler, but effective.
}

