% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.listTimezones}
\alias{h2o.listTimezones}
\title{List all of the Time Zones Acceptable by the H2O cluster.}
\usage{
h2o.listTimezones()
}
\description{
List all of the Time Zones Acceptable by the H2O cluster.
}
