% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.log}
\alias{h2o.log}
\title{Compute the logarithm of x}
\usage{
h2o.log(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the logarithm of x
}
\seealso{
\code{\link[base]{log}} for the base R implementation.
}

