% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.log1p}
\alias{h2o.log1p}
\title{Compute the log1p of x}
\usage{
h2o.log1p(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the log1p of x
}
\seealso{
\code{\link[base]{log1p}} for the base R implementation.
}

