% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.log2}
\alias{h2o.log2}
\title{Compute the log2 of x}
\usage{
h2o.log2(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the log2 of x
}
\seealso{
\code{\link[base]{log2}} for the base R implementation.
}

