% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logging.R
\name{h2o.logAndEcho}
\alias{h2o.logAndEcho}
\title{Log a message on the server-side logs}
\usage{
h2o.logAndEcho(message)
}
\arguments{
\item{message}{A character string with the message to write to the log.}
}
\description{
This is helpful when running several pieces of work one after the other on a single H2O
cluster and you want to make a notation in the H2O server side log where one piece of
work ends and the next piece of work begins.
}
\details{
\code{h2o.logAndEcho} sends a message to H2O for logging. Generally used for debugging purposes.
}

