% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.logloss}
\alias{h2o.logloss}
\title{Retrieve the Log Loss Value}
\usage{
h2o.logloss(object, valid = FALSE, ...)
}
\arguments{
\item{object}{a \linkS4class{H2OModelMetrics} object of the correct type.}

\item{valid}{Retreive the validation metric.}

\item{\dots}{Extra arguments to be passed if `object` is of type
\linkS4class{H2OModel} (e.g. train=TRUE)}
}
\description{
Retrieves the log loss output for a \linkS4class{H2OBinomialMetrics} or
\linkS4class{H2OMultinomialMetrics} object
}

