% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.ls}
\alias{h2o.ls}
\title{List Keys on an H2O Cluster}
\usage{
h2o.ls()
}
\value{
Returns a list of hex keys in the current H2O instance.
}
\description{
Accesses a list of object keys in the running instance of H2O.
}
\examples{
\donttest{
library(h2o)
h2o.init()
prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)
h2o.ls()
}
}

