% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.lstrip}
\alias{h2o.lstrip}
\title{Strip set from left}
\usage{
h2o.lstrip(x, set = " ")
}
\arguments{
\item{x}{The column whose strings should be lstrip-ed.}

\item{set}{string of characters to be removed}
}
\description{
Return a copy of the target column with leading characters removed. The set argument
is a string specifying the set of characters to be removed. If omitted, the set
argument defaults to removing whitespace.
}
\examples{
\donttest{
library(h2o)
h2o.init()
string_to_lstrip <- as.h2o("1234567890")
lstrip_string <- h2o.lstrip(string_to_lstrip,"123") #Remove "123"
}
}

