% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.makeGLMModel}
\alias{h2o.makeGLMModel}
\title{Remake an H2O GLM Model}
\usage{
h2o.makeGLMModel(model, beta)
}
\arguments{
\item{model}{an \linkS4class{H2OModel} corresponding from a \code{h2o.glm} call.}

\item{beta}{a new set of beta_constraints}
}
\description{
This function allows the usage of new beta constraints to create an GLM model, from an existing
model.
}

