% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.max}
\alias{h2o.max}
\title{Returns the maxima of the input values.}
\usage{
h2o.max(x, na.rm = FALSE)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{\code{logical}. indicating whether missing values should be removed.}
}
\description{
Returns the maxima of the input values.
}
\seealso{
\code{\link[base]{max}} for the base R implementation.
}
