% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.min}
\alias{h2o.min}
\title{Returns the minima of the input values.}
\usage{
h2o.min(x, na.rm = FALSE)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{\code{logical}. indicating whether missing values should be removed.}
}
\description{
Returns the minima of the input values.
}
\seealso{
\code{\link[base]{min}} for the base R implementation.
}

