% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.mktime}
\alias{h2o.mktime}
\title{Compute msec since the Unix Epoch}
\usage{
h2o.mktime(
  year = 1970,
  month = 0,
  day = 0,
  hour = 0,
  minute = 0,
  second = 0,
  msec = 0
)
}
\arguments{
\item{year}{Defaults to 1970}

\item{month}{zero based (months are 0 to 11)}

\item{day}{zero based (days are 0 to 30)}

\item{hour}{hour}

\item{minute}{minute}

\item{second}{second}

\item{msec}{msec}
}
\description{
Compute msec since the Unix Epoch
}
