% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.month}
\alias{h2o.month}
\alias{month}
\alias{month.Frame}
\title{Convert Milliseconds to Months in H2O Datasets}
\usage{
h2o.month(x)

month(x)

\method{month}{Frame}(x)
}
\arguments{
\item{x}{An H2O Frame object.}
}
\value{
A Frame object containing the entries of \code{x} converted to months of
        the year.
}
\description{
Converts the entries of a Frame object from milliseconds to months (on a 1 to
12 scale).
}
\seealso{
\code{\link{h2o.year}}
}

