% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.na_omit}
\alias{h2o.na_omit}
\title{Remove Rows With NAs}
\usage{
h2o.na_omit(object, ...)
}
\arguments{
\item{object}{H2OFrame object}

\item{...}{Ignored}
}
\value{
Returns an H2OFrame object containing non-NA rows.
}
\description{
Remove Rows With NAs
}
