% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.names}
\alias{h2o.names}
\title{Column names of an H2OFrame}
\usage{
h2o.names(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Column names of an H2OFrame
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "http://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.names(iris)
}
}
\seealso{
\code{\link[base]{names}} for the base R implementation.
}
