% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.nchar}
\alias{h2o.nchar}
\title{String length}
\usage{
h2o.nchar(x)
}
\arguments{
\item{x}{The column whose string lengths will be returned.}
}
\description{
String length
}
\examples{
\dontrun{
library(h2o)
h2o.init()
string_to_nchar <- as.h2o("r tutorial")
nchar_string <- h2o.nchar(string_to_nchar)
}
}
