% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.networkTest}
\alias{h2o.networkTest}
\title{View Network Traffic Speed}
\usage{
h2o.networkTest(conn = h2o.getConnection())
}
\arguments{
\item{conn}{an \linkS4class{H2OConnection} object.}
}
\value{
Returns a table listing the network speed for 1B, 10KB, and 10MB.
}
\description{
View speed with various file sizes.
}

