% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.nlevels}
\alias{h2o.nlevels}
\title{Get the number of factor levels for this frame.}
\usage{
h2o.nlevels(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Get the number of factor levels for this frame.
}
\seealso{
\code{\link[base]{nlevels}} for the base R method.
}

