% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.nrow}
\alias{h2o.nrow}
\title{Return the number of rows present in x.}
\usage{
h2o.nrow(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Return the number of rows present in x.
}
\seealso{
\code{\link[base]{nrow}} for the base R implementation.
}
