% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.num_iterations}
\alias{h2o.num_iterations}
\title{Retrieve the number of iterations.}
\usage{
h2o.num_iterations(object)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}

\item{\dots}{further arguments to be passed on (currently unimplemented)}
}
\description{
Retrieve the number of iterations.
}
