% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.num_valid_substrings}
\alias{h2o.num_valid_substrings}
\title{Count of substrings >= 2 chars that are contained in file}
\usage{
h2o.num_valid_substrings(x, path)
}
\arguments{
\item{x}{The column on which to calculate the number of valid substrings.}

\item{path}{Path to text file containing line-separated strings to be referenced.}
}
\description{
Find the count of all possible substrings >= 2 chars that are contained in the specified line-separated text file.
}

