% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{h2o.openLog}
\alias{h2o.openLog}
\title{View H2O R Logs}
\usage{
h2o.openLog(type)
}
\arguments{
\item{type}{Currently unimplemented.}
}
\description{
Open existing logs of H2O R POST commands and error resposnes on local disk.
Used primarily for debugging purposes.
}
\examples{
\dontrun{
h2o.init()

h2o.startLogging()
australia_path = system.file("extdata", "australia.csv", package = "h2o")
australia = h2o.importFile(path = australia_path)
h2o.stopLogging()

# Not run to avoid windows being opened during R CMD check
# h2o.openLog("Command")
# h2o.openLog("Error")
}
}
\seealso{
\code{\link{h2o.startLogging}, \link{h2o.stopLogging},
         \link{h2o.clearLog}}
}
