% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.predict}
\alias{h2o.predict}
\title{Predict on an H2O Model}
\usage{
h2o.predict(object, newdata, ...)
}
\arguments{
\item{object}{a fitted model object for which prediction is desired.}

\item{newdata}{An H2OFrame object in which to look for
variables with which to predict.}

\item{...}{additional arguments to pass on.}
}
\value{
Returns an H2OFrame object with probabilites and
        default predictions.
}
\description{
Predict on an H2O Model
}
