% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.predicted_vs_actual_by_variable}
\alias{h2o.predicted_vs_actual_by_variable}
\title{Calculates per-level mean of predicted value vs actual value for a given variable.}
\usage{
h2o.predicted_vs_actual_by_variable(object, newdata, predicted, variable)
}
\arguments{
\item{object}{A trained supervised H2O model.}

\item{newdata}{Input frame (can be training/test/.. frame).}

\item{predicted}{Frame of predictions for the given input frame.}

\item{variable}{Name of variable to inspect.}
}
\value{
H2OTable
}
\description{
In the basic setting, this function is equivalent to doing group-by on variable and calculating
mean on predicted and actual. In addition to that it also handles NAs in response and weights
automatically.
}
