% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.pro_substrings_words}
\alias{h2o.pro_substrings_words}
\title{Proportion of substrings >= 2 chars that are contained in file}
\usage{
h2o.pro_substrings_words(x, path)
}
\arguments{
\item{x}{The column on which to calculate the proportion of valid substrings.}

\item{path}{Path to text file containing line-separated strings to be referenced.}
}
\description{
Find the proportion of all possible substrings >= 2 chars that are contained in the specified line-separated text file.
}

