% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.prod}
\alias{h2o.prod}
\title{Return the product of all the values present in its arguments.}
\usage{
h2o.prod(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Return the product of all the values present in its arguments.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.prod(iris["petal_len"])
}
}
\seealso{
\code{\link[base]{prod}} for the base R implementation.
}
