% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.range}
\alias{h2o.range}
\title{Returns a vector containing the minimum and maximum of all the given arguments.}
\usage{
h2o.range(x, na.rm = FALSE, finite = FALSE)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{\code{logical}. indicating whether missing values should be removed.}

\item{finite}{\code{logical}. indicating if all non-finite elements should be omitted.}
}
\description{
Returns a vector containing the minimum and maximum of all the given arguments.
}
\seealso{
\code{\link[base]{range}} for the base R implementation.
}

