% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvstore.R
\name{h2o.rapids}
\alias{h2o.rapids}
\title{Execute a Rapids expression.}
\usage{
h2o.rapids(expr)
}
\arguments{
\item{expr}{The rapids expression (ascii string)}
}
\description{
Execute a Rapids expression.
}
\examples{
\dontrun{
h2o.rapids('(setproperty "sys.ai.h2o.algos.evaluate_auto_model_parameters" "true")')
}
}
