% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.relevel}
\alias{h2o.relevel}
\title{Reorders levels of an H2O factor, similarly to standard R's relevel.}
\usage{
h2o.relevel(x, y)
}
\arguments{
\item{x}{factor column in h2o frame}

\item{y}{reference level (string)}
}
\value{
new reordered factor column
}
\description{
The levels of a factor are reordered os that the reference level is at level 0, remaining levels are moved down as needed.
}

