% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.removeVecs}
\alias{h2o.removeVecs}
\title{Delete Columns from a Frame}
\usage{
h2o.removeVecs(data, cols)
}
\arguments{
\item{data}{The Frame.}

\item{cols}{The columns to remove.}
}
\description{
Delete the specified columns from the Frame.  Returns a Frame without the specified
columns.
}

