% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.rep_len}
\alias{h2o.rep_len}
\title{Replicate Elements of Vectors or Lists into H2O}
\usage{
h2o.rep_len(x, length.out)
}
\arguments{
\item{x}{an H2O frame}

\item{length.out}{non negative integer. The desired length of the output
vector.}
}
\value{
Creates an H2OFrame of the same type as x
}
\description{
\code{h2o.rep_len} performs just as \code{rep} does. It replicates the values in
\code{x} in the H2O backend.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "http://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.rep_len(iris, length.out = 3)
}
}
