% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.residual_deviance}
\alias{h2o.residual_deviance}
\title{Retrieve the residual deviance}
\usage{
h2o.residual_deviance(object, train = FALSE, valid = FALSE, xval = FALSE)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}}

\item{train}{Retrieve the training residual deviance}

\item{valid}{Retrieve the validation residual deviance}

\item{xval}{Retrieve the cross-validation residual deviance}
}
\description{
If "train", "valid", and "xval" parameters are FALSE (default), then the training residual deviance value is returned. If more
than one parameter is set to TRUE, then a named vector of residual deviances are returned, where the names are "train", "valid"
or "xval".
}
