% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.residual_dof}
\alias{h2o.residual_dof}
\title{Retrieve the residual degrees of freedom
If "train", "valid", and "xval" parameters are FALSE (default), then the training residual degrees of freedom value is returned. If more
than one parameter is set to TRUE, then a named vector of residual degrees of freedom are returned, where the names are "train", "valid"
or "xval".}
\usage{
h2o.residual_dof(object, train = FALSE, valid = FALSE, xval = FALSE)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}}

\item{train}{Retrieve the training residual degrees of freedom}

\item{valid}{Retrieve the validation residual degrees of freedom}

\item{xval}{Retrieve the cross-validation residual degrees of freedom}
}
\description{
Retrieve the residual degrees of freedom
If "train", "valid", and "xval" parameters are FALSE (default), then the training residual degrees of freedom value is returned. If more
than one parameter is set to TRUE, then a named vector of residual degrees of freedom are returned, where the names are "train", "valid"
or "xval".
}

