% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.result}
\alias{h2o.result}
\title{Retrieve the results to view the best predictor subsets.}
\usage{
h2o.result(model)
}
\arguments{
\item{model}{H2OModelSelection  object}
}
\value{
Returns an H2OFrame object
}
\description{
Retrieve the results to view the best predictor subsets.
}
