% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{h2o.resume}
\alias{h2o.resume}
\title{Triggers auto-recovery resume - this will look into configured recovery dir and resume and
tasks that were interrupted by unexpected cluster stopping.}
\usage{
h2o.resume(recovery_dir = NULL)
}
\arguments{
\item{recovery_dir}{A \code{character} path to where cluster recovery data is stored, if blank, will use
cluster's configuration.}
}
\description{
Triggers auto-recovery resume - this will look into configured recovery dir and resume and
tasks that were interrupted by unexpected cluster stopping.
}
