% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{h2o.resumeGrid}
\alias{h2o.resumeGrid}
\title{Resume previously stopped grid training.}
\usage{
h2o.resumeGrid(grid_id, recovery_dir = NULL, ...)
}
\arguments{
\item{grid_id}{ID of existing grid search}

\item{recovery_dir}{When specified the grid and all necessary data (frames, models) will be saved to this
directory (use HDFS or other distributed file-system). Should the cluster crash during training, the grid
can be reloaded from this directory via \code{h2o.loadGrid} and training can be resumed}

\item{...}{Additional parameters to modify the resumed Grid.}
}
\description{
Resume previously stopped grid training.
}
