% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.rm}
\alias{h2o.rm}
\title{Delete Objects In H2O}
\usage{
h2o.rm(ids)
}
\arguments{
\item{ids}{The hex key associated with the object to be removed.}
}
\description{
Remove the h2o Big Data object(s) having the key name(s) from ids.
}
\seealso{
\code{\link{h2o.assign}}, \code{\link{h2o.ls}}
}

