% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.rstrip}
\alias{h2o.rstrip}
\title{Strip set from right}
\usage{
h2o.rstrip(x, set = " ")
}
\arguments{
\item{x}{The column whose strings should be rstrip-ed.}

\item{set}{string of characters to be removed}
}
\description{
Return a copy of the target column with leading characters removed. The set argument
is a string specifying the set of characters to be removed. If omitted, the set
argument defaults to removing whitespace.
}

