% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulefit.R
\name{h2o.rule_importance}
\alias{h2o.rule_importance}
\title{This function returns the table with estimated coefficients and language representations (in case it is a rule) 
for each of the significant baselearners.}
\usage{
h2o.rule_importance(model)
}
\arguments{
\item{model}{of the interest}
}
\description{
This function returns the table with estimated coefficients and language representations (in case it is a rule) 
for each of the significant baselearners.
}
