% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.scoreHistory}
\alias{h2o.scoreHistory}
\title{Retrieve Model Score History}
\usage{
h2o.scoreHistory(object)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}
}
\description{
Retrieve Model Score History
}

