% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.scoreHistoryGAM}
\alias{h2o.scoreHistoryGAM}
\title{Retrieve GLM Model Score History buried in GAM model}
\usage{
h2o.scoreHistoryGAM(object)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}
}
\description{
Retrieve GLM Model Score History buried in GAM model
}
