% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{h2o.screeplot}
\alias{h2o.screeplot}
\title{Scree Plot}
\usage{
h2o.screeplot(model, type = c("barplot", "lines"))
}
\arguments{
\item{model}{A PCA model}

\item{type}{Type of the plot. Either "barplot" or "lines".}
}
\description{
Scree Plot
}
