% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h2o.set_s3_credentials}
\alias{h2o.set_s3_credentials}
\title{Creates a new Amazon S3 client internally with specified credentials.}
\usage{
h2o.set_s3_credentials(secretKeyId, secretAccessKey)
}
\arguments{
\item{secretKeyId}{Amazon S3 Secret Key ID (provided by Amazon)}

\item{secretAccessKey}{Amazon S3 Secret Access Key (provided by Amazon)}
}
\description{
There are no validations done to the credentials. Incorrect credentials are thus revealed with first S3 import call.
}
