% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.setTimezone}
\alias{h2o.setTimezone}
\title{Set the Time Zone on the H2O Cloud}
\usage{
h2o.setTimezone(tz, conn = h2o.getConnection())
}
\arguments{
\item{tz}{The desired timezone.}

\item{conn}{An H2OConnection object.}
}
\description{
Set the Time Zone on the H2O Cloud
}

