% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.sqrt}
\alias{h2o.sqrt}
\title{Compute the square root of x}
\usage{
h2o.sqrt(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the square root of x
}
\seealso{
\code{\link[base]{sqrt}} for the base R implementation.
}

