% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.startLogging}
\alias{h2o.startLogging}
\title{Start Writing H2O R Logs}
\usage{
h2o.startLogging(file)
}
\arguments{
\item{file}{a character string name for the file, automatically generated}
}
\description{
Begin logging H2o R POST commands and error responses to local disk. Used
primarily for debuggin purposes.
}
\examples{
\donttest{
library(h2o)
localH2O = h2o.init()
h2o.startLogging()
ausPath = system.file("extdata", "australia.csv", package="h2o")
australia.hex = h2o.importFile(localH2O, path = ausPath)
h2o.stopLogging()
}
}
\seealso{
\code{\link{h2o.stopLogging}, \link{h2o.clearLog},
         \link{h2o.openLog}}
}

