% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{h2o.stopLogging}
\alias{h2o.stopLogging}
\title{Stop Writing H2O R Logs}
\usage{
h2o.stopLogging()
}
\description{
Halt logging of H2O R POST commands and error responses to local disk. Used
primarily for debugging purposes.
}
\examples{
\donttest{
library(h2o)
h2o.init()
h2o.startLogging()
ausPath = system.file("extdata", "australia.csv", package="h2o")
australia.hex = h2o.importFile(path = ausPath)
h2o.stopLogging()
}
}
\seealso{
\code{\link{h2o.startLogging}, \link{h2o.clearLog},
         \link{h2o.openLog}}
}
