% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.str}
\alias{h2o.str}
\title{Display the structure of an H2OFrame object}
\usage{
h2o.str(object, ..., cols = FALSE)
}
\arguments{
\item{object}{An H2OFrame.}

\item{...}{Further arguments to be passed from or to other methods.}

\item{cols}{Print the per-column str for the H2OFrame}
}
\description{
Display the structure of an H2OFrame object
}

