% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.strsplit}
\alias{h2o.strsplit}
\title{String Split}
\usage{
h2o.strsplit(x, split)
}
\arguments{
\item{x}{The column whose strings must be split.}

\item{split}{The pattern to split on.}
}
\value{
An H2OFrame where each column is the outcome of the string split.
}
\description{
String Split
}
\examples{
\donttest{
library(h2o)
h2o.init()
string_to_split <- as.h2o("Split at every character.")
split_string <- h2o.strsplit(string_to_split,"")
}
}

