% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.sub}
\alias{h2o.sub}
\title{String Substitute}
\usage{
h2o.sub(pattern, replacement, x, ignore.case = FALSE)
}
\arguments{
\item{pattern}{The pattern to replace.}

\item{replacement}{The replacement pattern.}

\item{x}{The column on which to operate.}

\item{ignore.case}{Case sensitive or not}
}
\description{
Mutates the input. Changes the first occurence of pattern with replacement.
}

