% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.subset}
\alias{h2o.subset}
\title{Subsetting an H2O Frame}
\usage{
h2o.subset(x, subset, select, drop = FALSE, ...)
}
\arguments{
\item{x}{a \linkS4class{H2OFrame} to be subsetted}

\item{subset}{logical expression indicating elements or rows to keep}

\item{select}{expression, indicating columns to select from a data frame}

\item{drop}{passed on the the \code{[} indexing operator}

\item{\dots}{further arguments to be passed to or from other methods}
}
\description{
Returns a subset of an \linkS4class{H2OFrame} which meets conditions.
}
\seealso{
For the base \code{R} implementation see \code{\link[base]{subset}}
}

