% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.sum}
\alias{h2o.sum}
\title{Return the sum of all the values present in its arguments.}
\usage{
h2o.sum(x, na.rm = FALSE)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{\code{logical}. indicating whether missing values should be removed.}
}
\description{
Return the sum of all the values present in its arguments.
}
\seealso{
\code{\link[base]{sum}} for the base R implementation.
}

