% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.tan}
\alias{h2o.tan}
\title{Compute the tangent of x}
\usage{
h2o.tan(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the tangent of x
}
\seealso{
\code{\link[base]{tan}} for the base R implementation.
}
