% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.tanh}
\alias{h2o.tanh}
\title{Compute the hyperbolic tangent of x}
\usage{
h2o.tanh(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the hyperbolic tangent of x
}
\seealso{
\code{\link[base]{tanh}} for the base R implementation.
}

