% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.tolower}
\alias{h2o.tolower}
\title{Convert strings to lowercase}
\usage{
h2o.tolower(x)
}
\arguments{
\item{x}{An H2OFrame object whose strings should be lower cased}
}
\value{
An H2OFrame with all entries in lowercase format
}
\description{
Convert strings to lowercase
}
\examples{
\donttest{
library(h2o)
h2o.init()
string_to_lower <- as.h2o("ABCDE")
lowered_string <- h2o.tolower(string_to_lower)
}
}

