% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.topBottomN}
\alias{h2o.topBottomN}
\title{H2O topBottomN}
\usage{
h2o.topBottomN(x, column, nPercent, grabTopN)
}
\arguments{
\item{x}{an H2OFrame}

\item{column}{is a column name or column index to grab the top N percent value from}

\item{nPercent}{a top percentage values to grab}

\item{grabTopN}{if -1 grab bottom percentage, 1 grab top percentage}
}
\value{
An H2OFrame with 2 columns: first column is the original row indices, second column contains the values
}
\description{
topBottomN function will will grab the top N percent or botom N percent of values of a column and return it in a
 H2OFrame.
}
