% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.tot_withinss}
\alias{h2o.tot_withinss}
\title{Get the total within cluster sum of squares.}
\usage{
h2o.tot_withinss(object, valid = FALSE, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}

\item{valid}{Retreive the validation metric.}

\item{\dots}{further arguments to be passed on (currently unimplemented)}
}
\description{
Get the total within cluster sum of squares.
}

