% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.toupper}
\alias{h2o.toupper}
\title{Convert strings to uppercase}
\usage{
h2o.toupper(x)
}
\arguments{
\item{x}{An H2OFrame object whose strings should be upper cased}
}
\value{
An H2OFrame with all entries in uppercase format
}
\description{
Convert strings to uppercase
}
\examples{
\donttest{
library(h2o)
h2o.init()
string_to_upper <- as.h2o("abcde")
upper_string <- h2o.toupper(string_to_upper)
}
}

