% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.transform}
\alias{h2o.transform}
\title{Use H2O Transformation model and apply the underlying transformation}
\usage{
h2o.transform(model, ...)
}
\arguments{
\item{model}{A trained model representing the transformation strategy}

\item{...}{Transformation model-specific parameters}
}
\value{
Returns an H2OFrame object with data transformed.
}
\description{
Use H2O Transformation model and apply the underlying transformation
}
