% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.unique}
\alias{h2o.unique}
\title{H2O Unique}
\usage{
h2o.unique(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\value{
Returns an H2OFrame object.
}
\description{
Extract unique values in the column.
}
