% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{h2o.upload_model}
\alias{h2o.upload_model}
\title{Upload a binary model from the provided local path to the H2O cluster.
(H2O model can be saved in a binary form either by saveModel() or by download_model() function.)}
\usage{
h2o.upload_model(path)
}
\arguments{
\item{path}{A path on the machine this python session is currently connected to, specifying the location of the model to upload.}
}
\value{
Returns a new \linkS4class{H2OModel} object.
}
\description{
Upload a binary model from the provided local path to the H2O cluster.
(H2O model can be saved in a binary form either by saveModel() or by download_model() function.)
}
\seealso{
\code{\link{h2o.saveModel}}, \code{\link{h2o.download_model}}
}
