% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.which}
\alias{h2o.which}
\title{Which indices are TRUE?}
\usage{
h2o.which(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\value{
Returns an H2OFrame object.
}
\description{
Give the TRUE indices of a logical object, allowing for array indices.
}
\examples{
\dontrun{
h2o.init()
iris_hf <- as.h2o(iris)
h2o.which(iris_hf[, 1] == 4.4)
}
}
\seealso{
\code{\link[base]{which}} for the base R method.
}
