% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.which_max}
\alias{h2o.which_max}
\alias{which.max.H2OFrame}
\alias{which.min.H2OFrame}
\title{Which indice contains the max value?}
\usage{
h2o.which_max(x, na.rm = TRUE, axis = 0)

which.max.H2OFrame(x, na.rm = TRUE, axis = 0)

which.min.H2OFrame(x, na.rm = TRUE, axis = 0)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{\code{logical}. Indicate whether missing values should be removed.}

\item{axis}{\code{integer}. Indicate whether to calculate the mean down a column (0) or across a row (1).}
}
\value{
Returns an H2OFrame object.
}
\description{
Get the index of the max value in a column or row
}
\seealso{
\code{\link[base]{which.max}} for the base R method.
}

