% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.which_min}
\alias{h2o.which_min}
\title{Which index contains the min value?}
\usage{
h2o.which_min(x, na.rm = TRUE, axis = 0)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{\code{logical}. Indicate whether missing values should be removed.}

\item{axis}{\code{integer}. Indicate whether to calculate the mean down a column (0) or across a row (1).}
}
\value{
Returns an H2OFrame object.
}
\description{
Get the index of the min value in a column or row
}
\seealso{
\code{\link[base]{which.min}} for the base R method.
}

