% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.year}
\alias{h2o.year}
\alias{year}
\alias{year.H2OFrame}
\title{Convert Milliseconds to Years in H2O Datasets}
\usage{
h2o.year(x)

year(x)

\method{year}{H2OFrame}(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\value{
An H2OFrame object containing the entries of \code{x} converted to years
}
\description{
Convert the entries of an H2OFrame object from milliseconds to years, indexed
starting from 1900.
}
\details{
This method calls the function of the MutableDateTime class in Java.
}
\seealso{
\code{\link{h2o.month}}
}

