% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{H2OConnectionMutableState}
\alias{H2OConnectionMutableState}
\title{The H2OConnectionMutableState class}
\description{
This class represents the mutable aspects of a connection to an H2O cluster.
}
\section{Slots}{

\describe{
\item{\code{session_id}}{A \code{character} string specifying the H2O session identifier.}

\item{\code{key_count}}{A \code{integer} value specifying count for the number of keys generated for the \code{session_id}.}
}}

