% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{H2OCoxPHModel-class}
\alias{H2OCoxPHModel}
\alias{H2OCoxPHModel-class}
\alias{show,H2OCoxPHModel-method}
\title{The H2OCoxPHModel object.}
\usage{
\S4method{show}{H2OCoxPHModel}(object)
}
\arguments{
\item{object}{an \code{H2OCoxPHModel} object.}
}
\description{
Virtual object representing H2O's CoxPH Model.
}

