% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{H2OCoxPHModelSummary-class}
\alias{H2OCoxPHModelSummary}
\alias{H2OCoxPHModelSummary-class}
\title{The H2OCoxPHModelSummary object.}
\description{
Wrapper object for summary information compatible with survival package.
}
\section{Slots}{

\describe{
\item{\code{summary}}{A \code{list} containing the a summary compatible with CoxPH summary used in the survival package.}
}}

