% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{H2OFrame-class}
\alias{H2OFrame}
\alias{H2OFrame-class}
\alias{show,H2OFrame-method}
\title{The H2OFrame class}
\usage{
\S4method{show}{H2OFrame}(object)
}
\arguments{
\item{object}{An \code{H2OConnection} object.}
}
\description{
The H2OFrame class
}
\section{Slots}{

\describe{
\item{\code{conn}}{An \code{H2OConnection} object specifying the connection to an H2O cloud.}

\item{\code{frame_id}}{A \code{character} string specifying the identifier for the frame in the H2O cloud.}

\item{\code{finalizers}}{A \code{list} object containing environments with finalizers that
remove objects from the H2O cloud.}

\item{\code{mutable}}{An \code{H2OFrameMutableState} object to hold the mutable state for the H2O frame.}
}}

