% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{H2OFrame-Extract}
\alias{$,H2OFrame-method}
\alias{$<-,H2OFrame-method}
\alias{H2OFrame-Extract}
\alias{[,H2OFrame-method}
\alias{[<-,H2OFrame-method}
\alias{[[,H2OFrame-method}
\alias{[[<-,H2OFrame-method}
\title{Extract or Replace Parts of an H2OFrame Object}
\usage{
\S4method{[}{H2OFrame}(x, i, j, ..., drop = TRUE)

\S4method{$}{H2OFrame}(x, name)

\S4method{[[}{H2OFrame}(x, i, exact = TRUE)

\S4method{[}{H2OFrame}(x, i, j, ...) <- value

\S4method{$}{H2OFrame}(x, name) <- value

\S4method{[[}{H2OFrame}(x, i) <- value
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i,j,...}{indices specifying elements to extract or replace. Indices are numeric or
character vectors or empty (missing) or will be matched to the names.}

\item{drop}{a logical, whether or not to attempt to reduce dimensions to the
lowest possible.}

\item{name}{a literal character string or a name (possibly backtick quoted).}

\item{exact}{controls possible partial matching of \code{[[} when extracting
a character}

\item{value}{an array-like H2O object similar to \code{x}.}
}
\description{
Operators to extract or replace parts of H2OFrame objects.
}

