% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{Ops.H2OFrame}
\alias{Ops.H2OFrame}
\alias{Math.H2OFrame}
\alias{Summary.H2OFrame}
\alias{!.H2OFrame}
\alias{is.na.H2OFrame}
\alias{t.H2OFrame}
\alias{log}
\alias{log10}
\alias{log2}
\alias{log1p}
\alias{trunc}
\alias{\%*\%}
\alias{nrow.H2OFrame}
\alias{ncol.H2OFrame}
\alias{length.H2OFrame}
\alias{h2o.length}
\alias{names<-.H2OFrame}
\alias{colnames<-}
\title{S3 Group Generic Functions for H2O}
\usage{
\method{Ops}{H2OFrame}(e1, e2)

\method{Math}{H2OFrame}(x, ...)

\method{Math}{H2OFrame}(x, ...)

\method{Math}{H2OFrame}(x, ...)

\method{Summary}{H2OFrame}(x, ..., na.rm)

\method{!}{H2OFrame}(x)

\method{is.na}{H2OFrame}(x)

\method{t}{H2OFrame}(x)

log(x, ...)

log10(x)

log2(x)

log1p(x)

trunc(x, ...)

x \%*\% y

nrow.H2OFrame(x)

ncol.H2OFrame(x)

\method{length}{H2OFrame}(x)

h2o.length(x)

\method{names}{H2OFrame}(x) <- value

colnames(x) <- value
}
\arguments{
\item{e1}{object}

\item{e2}{object}

\item{x}{object}

\item{...}{Further arguments passed to or from other methods.}

\item{na.rm}{logical. whether or not missing values should be removed}

\item{y}{object}

\item{value}{To be assigned}
}
\description{
Methods for group generic functions and H2O objects.
}
