% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{H2OInfogram-class}
\alias{H2OInfogram-class}
\title{H2OInfogram class}
\description{
H2OInfogram class contains a subset of what a normal H2OModel will return
}
\section{Slots}{

\describe{
\item{\code{model_id}}{string returned as part of every \code{H2OModel}}

\item{\code{algorithm}}{string denoting the algorithm used to build infogram}

\item{\code{admissible_features}}{string array denoting all predictor names which pass the cmi and relelvance threshold}

\item{\code{admissible_features_valid}}{string array denoting all predictor names which pass the cmi and relelvance threshold from validation frame}

\item{\code{admissible_features_xval}}{string array denoting all predictor names which pass the cmi and relelvance threshold from cv holdout set}

\item{\code{net_information_threshold}}{numeric value denoting threshold used for predictor selection}

\item{\code{total_information_threshold}}{numeric value denoting threshold used for predictor selection}

\item{\code{safety_index_threshold}}{numeric value denoting threshold used for predictor selection}

\item{\code{relevance_index_threshold}}{numeric value denoting threshold used for predictor selection}

\item{\code{admissible_score}}{\code{H2OFrame} that contains columns, admissible, admissible_index, relevance, cmi, cmi_raw}

\item{\code{admissible_score_valid}}{\code{H2OFrame} that contains columns, admissible, admissible_index, relevance, cmi, cmi_raw from validation frame}

\item{\code{admissible_score_xval}}{\code{H2OFrame} that contains averages of columns, admissible, admissible_index, relevance, cmi, cmi_raw from cv hold-out}
}}

