% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{H2OInfogram}
\alias{H2OInfogram}
\title{wrapper function for instantiating H2OInfogram}
\usage{
H2OInfogram(model_id, ...)
}
\arguments{
\item{model_id}{is string of H2OModel object}

\item{...}{parameters to algorithm, admissible_features, ...}
}
\value{
A \code{H2OInfogram} object
}
\description{
wrapper function for instantiating H2OInfogram
}
