% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{H2OLeafNode-class}
\alias{H2OLeafNode-class}
\title{The H2OLeafNode class.}
\description{
This class represents a single leaf node in an \code{H2OTree}.
}
\details{
#' @aliases H2OLeafNode
}
