% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{H2OModelMetrics-class}
\alias{H2OAutoEncoderMetrics-class}
\alias{H2OBinomialMetrics-class}
\alias{H2OClusteringMetrics-class}
\alias{H2ODimReductionMetrics-class}
\alias{H2OModelMetrics}
\alias{H2OModelMetrics-class}
\alias{H2OMultinomialMetrics-class}
\alias{H2ORegressionMetrics-class}
\alias{H2OUnknownMetrics-class}
\alias{show,H2OAutoEncoderMetrics-method}
\alias{show,H2OBinomialMetrics-method}
\alias{show,H2OClusteringMetrics-method}
\alias{show,H2ODimReductionMetrics-method}
\alias{show,H2OModelMetrics-method}
\alias{show,H2OMultinomialMetrics-method}
\alias{show,H2ORegressionMetrics-method}
\title{The H2OModelMetrics Object.}
\usage{
\S4method{show}{H2OModelMetrics}(object)

\S4method{show}{H2OBinomialMetrics}(object)

\S4method{show}{H2OMultinomialMetrics}(object)

\S4method{show}{H2ORegressionMetrics}(object)

\S4method{show}{H2OClusteringMetrics}(object)

\S4method{show}{H2OAutoEncoderMetrics}(object)

\S4method{show}{H2ODimReductionMetrics}(object)
}
\arguments{
\item{object}{An \code{H2OModelMetrics} object}
}
\description{
A class for constructing performance measures of H2O models.
}

