% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{H2ONode-class}
\alias{H2ONode-class}
\alias{show,H2ONode-method}
\title{The H2ONode class.}
\usage{
\S4method{show}{H2ONode}(object)
}
\arguments{
\item{object}{an \code{H2ONode} object.}
}
\description{
The H2ONode class.
}
\section{Slots}{

\describe{
\item{\code{id}}{An \code{integer} representing node's unique identifier. Generated by H2O.}

\item{\code{levels}}{A \code{character} representing categorical levels on split from parent's node belonging into this node. NULL for root node or non-categorical splits.

#' @aliases H2ONode}
}}

