% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{H2OObject-class}
\alias{H2OObject}
\alias{H2OObject-class}
\alias{initialize,H2OObject-method}
\title{The H2OObject class}
\usage{
\S4method{initialize}{H2OObject}(.Object, ...)
}
\arguments{
\item{.Object}{an \code{H2OObject}}

\item{\dots}{additional parameters to pass on to functions}
}
\description{
The H2OObject class
}
\section{Slots}{

\describe{
\item{\code{conn}}{An \code{H2OConnection} object specifying the connection to an H2O cloud.}

\item{\code{id}}{A \code{character} string specifying the key in the H2O cloud's key-value store.}

\item{\code{finalizers}}{A \code{list} object containing environments with finalizers that
remove keys from the H2O key-value store.}
}}

