% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{H2OS4groupGeneric}
\alias{!,H2OFrame-method}
\alias{H2OS4groupGeneric}
\alias{Math,H2OFrame-method}
\alias{Math2,H2OFrame-method}
\alias{Ops,H2OFrame,H2OFrame-method}
\alias{Ops,H2OFrame,character-method}
\alias{Ops,H2OFrame,missing-method}
\alias{Ops,H2OFrame,numeric-method}
\alias{Ops,character,H2OFrame-method}
\alias{Ops,missing,H2OFrame-method}
\alias{Ops,numeric,H2OFrame-method}
\alias{Summary,H2OFrame-method}
\alias{\%*\%,H2OFrame,H2OFrame-method}
\alias{is.na,H2OFrame-method}
\alias{log,H2OFrame-method}
\alias{t,H2OFrame-method}
\alias{trunc,H2OFrame-method}
\title{S4 Group Generic Functions for H2O}
\usage{
\S4method{Ops}{missing,H2OFrame}(e1, e2)

\S4method{Ops}{H2OFrame,missing}(e1, e2)

\S4method{Ops}{H2OFrame,H2OFrame}(e1, e2)

\S4method{Ops}{numeric,H2OFrame}(e1, e2)

\S4method{Ops}{H2OFrame,numeric}(e1, e2)

\S4method{Ops}{H2OFrame,character}(e1, e2)

\S4method{Ops}{character,H2OFrame}(e1, e2)

\S4method{Math}{H2OFrame}(x)

\S4method{Math2}{H2OFrame}(x, digits)

\S4method{Summary}{H2OFrame}(x, ..., na.rm = FALSE)

\S4method{!}{H2OFrame}(x)

\S4method{is.na}{H2OFrame}(x)

\S4method{t}{H2OFrame}(x)

\S4method{log}{H2OFrame}(x, ...)

\S4method{trunc}{H2OFrame}(x, ...)

\S4method{\%*\%}{H2OFrame,H2OFrame}(x, y)
}
\arguments{
\item{x,y,e1,e2}{objects.}

\item{digits}{number of digits to be used in \code{round} or \code{signif}}

\item{na.rm}{logical: should missing values be removed?}

\item{\dots}{further arguments passed to or from methods}
}
\description{
Methods for group generic functions and H2O objects.
}

