% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/segment.R
\docType{class}
\name{H2OSegmentModels-class}
\alias{H2OSegmentModels-class}
\alias{show,H2OSegmentModels-method}
\title{H2O Segment Models}
\usage{
\S4method{show}{H2OSegmentModels}(object)
}
\arguments{
\item{object}{an \code{H2OModel} object.}
}
\description{
A class to contain the information for segment models.
}
\section{Slots}{

\describe{
\item{\code{segment_models_id}}{the  identifier for the segment models collections}
}}

