% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{H2OSegmentModelsFuture-class}
\alias{H2OSegmentModelsFuture-class}
\title{H2O Future Segment Models}
\description{
A class to contain the information for background segment models jobs.
}
\section{Slots}{

\describe{
\item{\code{job_key}}{a character key representing the identification of the job process.}

\item{\code{segment_models_id}}{the final identifier for the segment models collections}
}}

\seealso{
\linkS4class{H2OSegmentModels} for the final segment models types.
}
