% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{class}
\name{H2OSplitNode-class}
\alias{H2OSplitNode-class}
\alias{H2OSplitNode}
\title{The H2OSplitNode class.}
\description{
This class represents a single non-terminal node in an \code{H2OTree}.
}
\section{Slots}{

\describe{
\item{\code{threshold}}{A \code{numeric} split threshold, typically when the split column is numerical.}

\item{\code{left_child}}{A \code{H2ONodeOrNULL} representing the left child node, if a node has one.}

\item{\code{right_child}}{A \code{H2ONodeOrNULL} representing the right child node, if a node has one.}

\item{\code{split_feature}}{A \code{character} representing the name of the column this node splits on.}

\item{\code{left_levels}}{A \code{character} representing the levels of a categorical feature heading to the left child of this node. NA for non-categorical split.}

\item{\code{right_levels}}{A \code{character} representing the levels of a categorical feature heading to the right child of this node. NA for non-categorical split.}

\item{\code{na_direction}}{A \code{character} representing the direction of NA values. LEFT means NA values go to the left child node, RIGH means NA values go to the right child node.}
}}

