% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{housevotes}
\alias{housevotes}
\title{United States Congressional Voting Records 1984}
\format{
A data frame with 435 rows and 17 columns
}
\source{
Congressional Quarterly Almanac, 98th Congress, 2nd session 1984, Volume XL: Congressional Quarterly Inc., Washington, D.C., 1985
}
\description{
This data set includes votes for each of the U.S. House of Representatives Congressmen on the 16 
key votes identified by the CQA. The CQA lists nine different types of votes: voted for, paired for, 
and announced for (these three simplified to yea), voted against, paired against, and announced 
against (these three simplified to nay), voted present, voted present to avoid conflict of interest, 
and did not vote or otherwise make a position known (these three simplified to an unknown disposition).
}
\references{
Newman, D.J. & Hettich, S. & Blake, C.L. & Merz, C.J. (1998). UCI Repository of machine 
learning databases [https://www.ics.uci.edu/~mlearn/MLRepository.html]. Irvine, CA: University of 
California, Department of Information and Computer Science.
}
\keyword{datasets}
