% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{initialize,H2OInfogram-method}
\alias{initialize,H2OInfogram-method}
\title{Method on \code{H2OInfogram} object which in this case is to instantiate and initialize it}
\usage{
\S4method{initialize}{H2OInfogram}(.Object, model_id, ...)
}
\arguments{
\item{.Object}{An \code{H2OInfogram} object}

\item{model_id}{string returned as part of every H2OModel}

\item{...}{additional arguments to pass on}
}
\value{
A \code{H2OInfogram} object
}
\description{
Method on \code{H2OInfogram} object which in this case is to instantiate and initialize it
}
