% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{iris}
\alias{iris}
\title{Edgar Anderson's Iris Data}
\format{
A data frame with 150 rows and 5 columns
}
\source{
Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems. Annals of Eugenics, 7, Part II, 179-188.

The data were collected by Anderson, Edgar (1935). The irises of the Gaspe Peninsula, Bulletin of the American Iris Society, 59, 2-5.
}
\description{
Measurements in centimeters of the sepal length and width and petal length and width, 
respectively, for three species of iris flowers.
}
\keyword{datasets}
