% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{is.character}
\alias{is.character}
\title{Check if character}
\usage{
is.character(x)
}
\arguments{
\item{x}{An H2OFrame object}
}
\description{
Check if character
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "http://s3.amazonaws.com/h2o-public-test-data/smalldata/coxph_test/heart.csv"
heart <- h2o.importFile(f)

heart["transplant"] <- as.character(heart["transplant"])
is.character(heart["transplant"])
}
}
