% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{is.factor}
\alias{is.factor}
\title{Check if factor}
\usage{
is.factor(x)
}
\arguments{
\item{x}{An H2OFrame object}
}
\description{
Check if factor
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/junit/cars_20mpg.csv"
cars <- h2o.importFile(f)
cars["economy_20mpg"] <- as.factor(cars["economy_20mpg"])
is.factor(cars["economy_20mpg"])
}
}
