% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{is.h2o}
\alias{is.h2o}
\title{Is H2O Frame object}
\usage{
is.h2o(x)
}
\arguments{
\item{x}{An \code{R} object.}
}
\description{
Test if object is H2O Frame.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

frame <- h2o.createFrame(rows = 6, cols = 2,
                         categorical_fraction = 0.0, 
                         missing_fraction = 0.7, 
                         seed = 123)
is.h2o(frame)
}
}
