% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{is.numeric,H2OFrame-method}
\alias{is.numeric,H2OFrame-method}
\title{Is H2O Data Frame column numeric}
\usage{
\S4method{is.numeric}{H2OFrame}(x)
}
\arguments{
\item{x}{an \linkS4class{H2OFrame} object column.}
}
\value{
Returns logical value.
}
\description{
Is H2O Data Frame column numeric
}

