% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\docType{methods}
\name{length,H2OTree-method}
\alias{length,H2OTree-method}
\title{Overrides the behavior of length() function on H2OTree class. Returns number of nodes in an \code{H2OTree}}
\usage{
\S4method{length}{H2OTree}(x)
}
\arguments{
\item{x}{An \code{H2OTree} to count nodes for.}
}
\description{
Overrides the behavior of length() function on H2OTree class. Returns number of nodes in an \code{H2OTree}
}

