% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{Logical-or}
\alias{Logical-or}
\alias{||}
\title{Logical or for H2OFrames}
\usage{
"||"(x, y)
}
\arguments{
\item{x}{An H2OFrame object}

\item{y}{An H2OFrame object}
}
\description{
Logical or for H2OFrames
}
