% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{median,H2OFrame-method}
\alias{median,H2OFrame-method}
\title{H2O Median}
\usage{
\S4method{median}{H2OFrame}(x, na.rm = TRUE)
}
\arguments{
\item{x}{An \linkS4class{H2OFrame} object.}

\item{na.rm}{a logical, indicating whether na's are omitted.}
}
\description{
Compute the airthmetic mean of a \linkS4class{H2OFrame}.
}
\examples{
localH2O <- h2o.init()
prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(localH2O, path = prosPath, destination_frame = "prostate.hex")
}

