% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ModelAccessors}
\alias{ModelAccessors}
\alias{getParms}
\alias{getParms,H2OModel-method}
\alias{getCenters}
\alias{getCentersStd}
\alias{getWithinSS}
\alias{getTotWithinSS}
\alias{getBetweenSS}
\alias{getTotSS}
\alias{getIterations}
\alias{getClusterSizes}
\alias{getCenters,H2OClusteringModel-method}
\alias{getCentersStd,H2OClusteringModel-method}
\alias{getWithinSS,H2OClusteringModel-method}
\alias{getTotWithinSS,H2OClusteringModel-method}
\alias{getBetweenSS,H2OClusteringModel-method}
\alias{getTotSS,H2OClusteringModel-method}
\alias{getIterations,H2OClusteringModel-method}
\alias{getClusterSizes,H2OClusteringModel-method}
\title{Accessor Methods for H2OModel Object}
\usage{
getParms(object)

\S4method{getParms}{H2OModel}(object)

getCenters(object)

getCentersStd(object)

getWithinSS(object)

getTotWithinSS(object)

getBetweenSS(object)

getTotSS(object)

getIterations(object)

getClusterSizes(object)

\S4method{getCenters}{H2OClusteringModel}(object)

\S4method{getCentersStd}{H2OClusteringModel}(object)

\S4method{getWithinSS}{H2OClusteringModel}(object)

\S4method{getTotWithinSS}{H2OClusteringModel}(object)

\S4method{getBetweenSS}{H2OClusteringModel}(object)

\S4method{getTotSS}{H2OClusteringModel}(object)

\S4method{getIterations}{H2OClusteringModel}(object)

\S4method{getClusterSizes}{H2OClusteringModel}(object)
}
\arguments{
\item{object}{an \linkS4class{H2OModel} class object.}
}
\description{
Function accessor methods for various H2O output fields.
}
