% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{na.omit,H2OFrame-method}
\alias{na.omit,H2OFrame-method}
\title{Remove Rows With NAs}
\usage{
\S4method{na.omit}{H2OFrame}(object, ...)
}
\arguments{
\item{object}{H2OFrame object}

\item{...}{Ignored}
}
\description{
Remove Rows With NAs
}

