% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{names.H2OFrame}
\alias{names.H2OFrame}
\title{Column names of an H2OFrame}
\usage{
\method{names}{H2OFrame}(x)
}
\arguments{
\item{x}{An H2OFrame}
}
\description{
Column names of an H2OFrame
}
\examples{
\dontrun{
library(h2o)
h2o.init()

frame <- h2o.createFrame(rows = 6, cols = 2,
                         categorical_fraction = 0.0, 
                         missing_fraction = 0.7, 
                         seed = 123)
names(frame)
}
}
