% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Node-class}
\alias{ASTApply-class}
\alias{ASTBody-class}
\alias{ASTElse-class}
\alias{ASTEmpty-class}
\alias{ASTFor-class}
\alias{ASTFun-class}
\alias{ASTIf-class}
\alias{ASTReturn-class}
\alias{ASTSeries-class}
\alias{ASTSpan-class}
\alias{Node}
\alias{Node-class}
\title{The Node class.}
\description{
An object of type Node inherits from an H2OFrame, but holds no H2O-aware data. Every node in the abstract syntax tree
has as its ancestor this class.

This class represents an operator between one or more H2O objects. ASTApply nodes are always root nodes in a tree and
are never leaf nodes. Operators are discussed more in depth in ops.R.
}
\details{
Every node in the abstract syntax tree will have a symbol table, which is a dictionary of types and names for
all the relevant variables and functions defined in the current scope. A missing symbol is therefore discovered
by looking up the tree to the nearest symbol table defining that symbol.
}

