% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{plot,H2OParetoFront-method}
\alias{plot,H2OParetoFront-method}
\title{Plot Pareto front}
\usage{
\S4method{plot}{H2OParetoFront}(x, y, ...)
}
\arguments{
\item{x}{\code{H2OParetoFront} object}

\item{y}{missing}

\item{...}{unused}
}
\description{
Plot Pareto front
}
