% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl.R
\name{predict.H2OAutoML}
\alias{predict.H2OAutoML}
\title{Predict on an AutoML object}
\usage{
\method{predict}{H2OAutoML}(object, newdata, ...)
}
\arguments{
\item{object}{a fitted \linkS4class{H2OAutoML} object for which prediction is
desired}

\item{newdata}{An H2OFrame object in which to look for
variables with which to predict.}

\item{...}{additional arguments to pass on.}
}
\value{
Returns an H2OFrame object with probabilites and
        default predictions.
}
\description{
Obtains predictions from an AutoML object.
}
\details{
This method generated predictions on the leader model from an AutoML run.
The order of the rows in the results is the same as the order in which the
data was loaded, even if some rows fail (for example, due to missing
values or unseen factor levels).
}
