% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.Frame}
\alias{print.Frame}
\title{Print An H2O Frame}
\usage{
\method{print}{Frame}(x, ...)
}
\arguments{
\item{x}{An H2O Frame object}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Print An H2O Frame
}

