% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{print.H2OTable}
\alias{print.H2OTable}
\title{Print method for H2OTable objects}
\usage{
\method{print}{H2OTable}(x, header = TRUE, ...)
}
\arguments{
\item{x}{An H2OTable object}

\item{header}{A logical value dictating whether or not the table name should be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The original x object
}
\description{
This will print a truncated view of the table if there are more than 20 rows.
}
