% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{prostate}
\alias{prostate}
\title{Prostate Cancer Study}
\format{A data frame with 380 rows and 9 columns}
\source{
Hosmer and Lemeshow (2000) Applied Logistic Regression: Second Edition.
}
\description{
Baseline exam results on prostate cancer patients from Dr. Donn Young at The Ohio State
University Comprehensive Cancer Center.
}
\keyword{datasets}

