% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{range,H2OFrame-method}
\alias{range,H2OFrame-method}
\title{Range of an H2O Column}
\usage{
\S4method{range}{H2OFrame}(x, na.rm = TRUE)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{ignore missing values}
}
\description{
Range of an H2O Column
}

