% Generated by roxygen2 (4.0.2): do not edit by hand
\name{range.H2OFrame}
\alias{range.H2OFrame}
\title{Range of an H2O Column}
\usage{
\method{range}{H2OFrame}(..., na.rm = TRUE)
}
\arguments{
\item{...}{An H2OFrame object.}

\item{na.rm}{ignore missing values}
}
\description{
Range of an H2O Column
}

