% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{sapply,H2OFrame-method}
\alias{sapply,H2OFrame-method}
\title{Apply Over a List in H2O}
\usage{
\S4method{sapply}{H2OFrame}(X, FUN, ..., simplify = TRUE, USE.NAMES = TRUE)
}
\arguments{
\item{X}{an \linkS4class{H2OFrame} object on which \code{apply} will operate.}

\item{FUN}{the function to be applied.}

\item{simplify,USE.NAMES}{ignored parameters from base funciton}

\item{\dots}{optional arguments to \code{FUN}.}
}
\description{
Functions equivalent to the default \code{R} sapply
}
\seealso{
\code{link[base]{sapply}} for the base implementation.
export
}

