% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{show,H2OAutoML-method}
\alias{show,H2OAutoML-method}
\title{Format AutoML object in user-friendly way}
\usage{
\S4method{show}{H2OAutoML}(object)
}
\arguments{
\item{object}{an \code{H2OAutoML} object.}
}
\description{
Format AutoML object in user-friendly way
}
