% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,H2OCoxPHModelSummary-method}
\alias{show,H2OCoxPHModelSummary-method}
\title{Print the CoxPH Model Summary}
\usage{
\S4method{show}{H2OCoxPHModelSummary}(object)
}
\arguments{
\item{object}{An \linkS4class{H2OCoxPHModelSummary} object.}

\item{...}{further arguments to be passed on (currently unimplemented)}
}
\description{
Print the CoxPH Model Summary
}
