% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str.Frame}
\alias{str.Frame}
\title{Display the structure of an H2O Frame object}
\usage{
\method{str}{Frame}(object, ..., cols = FALSE)
}
\arguments{
\item{object}{An H2O Frame.}

\item{...}{Further arguments to be passed from or to other methods.}

\item{cols}{Print the per-column str for the Frame}
}
\description{
Display the structure of an H2O Frame object
}

