% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{str.H2OFrame}
\alias{str.H2OFrame}
\title{Display the structure of an H2OFrame object}
\usage{
\method{str}{H2OFrame}(object, ..., cols = FALSE)
}
\arguments{
\item{object}{An H2OFrame.}

\item{...}{Further arguments to be passed from or to other methods.}

\item{cols}{Print the per-column str for the H2OFrame}
}
\description{
Display the structure of an H2OFrame object
}
