% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{summary,H2OCoxPHModel-method}
\alias{summary,H2OCoxPHModel-method}
\title{Summary method for H2OCoxPHModel objects}
\usage{
\S4method{summary}{H2OCoxPHModel}(object, conf.int = 0.95, scale = 1)
}
\arguments{
\item{object}{an \code{H2OCoxPHModel} object.}

\item{conf.int}{a specification of the confidence interval.}

\item{scale}{a scale.}
}
\description{
Summary method for H2OCoxPHModel objects
}
