% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary,H2OGrid-method}
\alias{summary,H2OGrid-method}
\title{Format grid object in user-friendly way}
\usage{
\S4method{summary}{H2OGrid}(object, show_stack_traces = F)
}
\arguments{
\item{object}{an \code{H2OGrid} object.}

\item{show_stack_traces}{a flag to show stack traces for model failures}
}
\description{
Format grid object in user-friendly way
}

