% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary,H2OModel-method}
\alias{summary,H2OModel-method}
\title{Print the Model Summary}
\usage{
\S4method{summary}{H2OModel}(object, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}

\item{...}{further arguments to be passed on (currently unimplemented)}
}
\description{
Print the Model Summary
}

