% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{walking}
\alias{walking}
\title{Muscular Actuations for Walking Subject}
\format{
A data frame with 151 rows and 124 columns
}
\description{
The musculoskeletal model, experimental data, settings files, and results 
for three-dimensional, muscle-actuated simulations at walking speed as 
described in Hamner and Delp (2013). Simulations were generated using OpenSim 2.4.
The data is available from \url{https://simtk.org/frs/index.php?group_id=603}.
}
\references{
Hamner, S.R., Delp, S.L. Muscle contributions to fore-aft and vertical body mass center accelerations over a range of running speeds. Journal of Biomechanics, vol 46, pp 780-787. (2013)
}
\keyword{datasets}
