% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{with_no_h2o_progress}
\alias{with_no_h2o_progress}
\title{Suppresses h2o progress output from \code{expr}}
\usage{
with_no_h2o_progress(expr)
}
\arguments{
\item{expr}{expression}
}
\value{
result of \code{expr}
}
\description{
Suppresses h2o progress output from \code{expr}
}
