% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.character.H2OFrame}
\alias{as.character.H2OFrame}
\title{Convert an H2OFrame to a String}
\usage{
\method{as.character}{H2OFrame}(x, ...)
}
\arguments{
\item{x}{An H2OFrame object}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Convert an H2OFrame to a String
}

