% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.factor}
\alias{as.factor}
\title{Convert H2O Data to Factors}
\usage{
as.factor(x)
}
\arguments{
\item{x}{a column from an H2OFrame data set.}
}
\description{
Convert a column into a factor column.
}
\examples{
\donttest{
h2o.init()
prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)
prostate.hex[,2] <- as.factor(prostate.hex[,2])
summary(prostate.hex)
}
}
\seealso{
\code{\link{is.factor}}.
}

