% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.h2o}
\alias{as.h2o}
\title{R data.frame -> Frame}
\usage{
as.h2o(x, destination_frame = "")
}
\arguments{
\item{x}{An \code{R} data frame.}

\item{destination_frame}{A string with the desired name for the H2O Frame.}
}
\description{
Import a local R data frame to the H2O cloud.
}

