% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{as.vector.H2OFrame}
\alias{as.vector.H2OFrame}
\title{Convert an H2OFrame to a vector}
\usage{
\method{as.vector}{H2OFrame}(x,mode)
}
\arguments{
\item{x}{An H2OFrame object}

\item{mode}{Mode to coerce vector to}
}
\description{
Convert an H2OFrame to a vector
}

