% Generated by roxygen2 (4.0.2): do not edit by hand
\name{colnames}
\alias{colnames}
\title{Returns the column names of an H2OFrame}
\usage{
colnames(x, do.NULL = TRUE, prefix = "col")
}
\arguments{
\item{x}{An H2OFrame object.}

\item{do.NULL}{logical. If FALSE and names are NULL, names are created.}

\item{prefix}{for created names.}
}
\description{
Returns the column names of an H2OFrame
}

