% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{colnames}
\alias{colnames}
\title{Returns the column names of an H2OFrame}
\usage{
colnames(x, do.NULL = TRUE, prefix = "col")
}
\arguments{
\item{x}{An H2OFrame object.}

\item{do.NULL}{logical. If FALSE and names are NULL, names are created.}

\item{prefix}{for created names.}
}
\description{
Returns the column names of an H2OFrame
}
\examples{
\donttest{
h2o.init()
iris.hex <- as.h2o(iris)
colnames(iris)  # Returns "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"
}
}

