% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{dimnames.H2OFrame}
\alias{dimnames.H2OFrame}
\title{Column names of an H2OFrame}
\usage{
\method{dimnames}{H2OFrame}(x)
}
\arguments{
\item{x}{An H2OFrame}
}
\description{
Column names of an H2OFrame
}

