% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.aic}
\alias{h2o.aic}
\title{Retrieve the AIC.}
\usage{
h2o.aic(object, valid = FALSE, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}.}

\item{valid}{Retrieve the validation AIC}

\item{\dots}{extra arguments to be passed if `object` is of type
\linkS4class{H2OModel} (e.g. train=TRUE)}
}
\description{
Retrieve the AIC.
}

