% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.aic}
\alias{h2o.aic}
\title{Retrieve the AIC.
If "train", "valid", and "xval" parameters are FALSE (default), then the training AIC value is returned. If more
than one parameter is set to TRUE, then a named vector of AICs are returned, where the names are "train", "valid"
or "xval".}
\usage{
h2o.aic(object, train = FALSE, valid = FALSE, xval = FALSE)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}.}

\item{train}{Retrieve the training AIC}

\item{valid}{Retrieve the validation AIC}

\item{xval}{Retrieve the cross-validation AIC}
}
\description{
Retrieve the AIC.
If "train", "valid", and "xval" parameters are FALSE (default), then the training AIC value is returned. If more
than one parameter is set to TRUE, then a named vector of AICs are returned, where the names are "train", "valid"
or "xval".
}

