% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.any}
\alias{h2o.any}
\title{Given a set of logical vectors, is at least one of the values true?}
\usage{
h2o.any(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Given a set of logical vectors, is at least one of the values true?
}
\seealso{
\code{\link[base]{all}} for the base R implementation.
}

