% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.any}
\alias{h2o.any}
\title{Given a set of logical vectors, is at least one of the values true?}
\usage{
h2o.any(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Given a set of logical vectors, is at least one of the values true?
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.any(iris[, 1] < 1000)
}
}
\seealso{
\code{\link[base]{all}} for the base R implementation.
}
