% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.anyFactor}
\alias{h2o.anyFactor}
\title{Check H2OFrame columns for factors}
\usage{
h2o.anyFactor(x)
}
\arguments{
\item{x}{An \code{\linkS4class{H2OFrame}} object.}
}
\value{
Returns a logical value indicating whether any of the columns in \code{x} are factors.
}
\description{
Determines if any column of an H2OFrame object contains categorical data.
}
\examples{
library(h2o)
localH2O <- h2o.init()
irisPath <- system.file("extdata", "iris_wheader.csv", package="h2o")
iris.hex <- h2o.importFile(localH2O, path = irisPath)
h2o.anyFactor(iris.hex)
}

