% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.arrange}
\alias{h2o.arrange}
\title{Sorts H2OFrame by the columns specified. H2OFrame should not contain any String columns.  Otherwise, an error will be thrown.  Returns a new H2OFrame, like dplyr::arrange.}
\usage{
h2o.arrange(x, ...)
}
\arguments{
\item{x}{The H2OFrame input to be sorted.}

\item{\dots}{The column names to sort by.}
}
\description{
Sorts H2OFrame by the columns specified. H2OFrame should not contain any String columns.  Otherwise, an error will be thrown.  Returns a new H2OFrame, like dplyr::arrange.
}

