% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.arrange}
\alias{h2o.arrange}
\title{Sorts an H2O frame by columns}
\usage{
h2o.arrange(x, ...)
}
\arguments{
\item{x}{The H2OFrame input to be sorted.}

\item{\dots}{The column names to sort by.}
}
\description{
Sorts H2OFrame by the columns specified. H2OFrame can contain String columns but should not sort on any
String columns.  Otherwise, an error will
 be thrown.  To sort column c1 in descending order, do desc(c1).  Returns a new H2OFrame, like dplyr::arrange.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.arrange(iris, "species","petal_len","petal_wid")
}

}
