% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.as_date}
\alias{h2o.as_date}
\title{Functions to convert between character representations and objects of class "Date" representing calendar dates.}
\usage{
h2o.as_date(x, format, ...)
}
\arguments{
\item{x}{H2OFrame column of strings or factors to be converted}

\item{format}{A character string indicating date pattern}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
Functions to convert between character representations and objects of class "Date" representing calendar dates.
}

