% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.ascharacter}
\alias{h2o.ascharacter}
\title{Convert H2O Data to Characters}
\usage{
h2o.ascharacter(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Convert H2O Data to Characters
}
\seealso{
\code{\link[base]{as.character}} for the base R implementation.
}

