% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.asfactor}
\alias{h2o.asfactor}
\title{Convert H2O Data to Factors}
\usage{
h2o.asfactor(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Convert H2O Data to Factors
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/junit/cars_20mpg.csv"
cars <- h2o.importFile(f)
h2o.asfactor(cars["cylinders"])
}
}
\seealso{
\code{\link[base]{as.numeric}} for the base R implementation.
}
