% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.asnumeric}
\alias{h2o.asnumeric}
\title{Convert H2O Data to Numerics}
\usage{
h2o.asnumeric(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Convert H2O Data to Numerics
}
\seealso{
\code{\link[base]{as.numeric}} for the base R implementation.
}

