% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.bottomN}
\alias{h2o.bottomN}
\title{H2O bottomN}
\usage{
h2o.bottomN(x, column, nPercent)
}
\arguments{
\item{x}{an H2OFrame}

\item{column}{is a column name or column index to grab the top N percent value from}

\item{nPercent}{is a bottom percentage value to grab}
}
\value{
An H2OFrame with 2 columns.  The first column is the original row indices, second column contains the bottomN values
}
\description{
bottomN function will will grab the bottom N percent of values of a column and return it in a H2OFrame.
Extract the top N percent of values of a column and return it in a H2OFrame.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f1 <- "https://s3.amazonaws.com/h2o-public-test-data/bigdata/laptop/jira/TopBottomNRep4.csv.zip"
f2 <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/jira/Bottom20Per.csv.zip"
data_Frame <- h2o.importFile(f1)
bottom_Answer <- h2o.importFile(f2)
nPercent <- c(1, 2, 3, 4)
frame_Names <- names(data_Frame)
nP <- nPercent[sample(1:length(nPercent), 1, replace = FALSE)]
col_Index <- sample(1:length(frame_Names), 1, replace = FALSE)
h2o.bottomN(data_Frame, frame_Names[col_Index], nP)
}
}
