% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.ceiling}
\alias{h2o.ceiling}
\title{Take a single numeric argument and return a numeric vector with the smallest integers}
\usage{
h2o.ceiling(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
ceiling takes a single numeric argument x and returns a
numeric vector containing the smallest integers not less than the
corresponding elements of x.
}
\seealso{
\code{\link[base]{ceiling}} for the base R implementation.
}

