% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.centers}
\alias{h2o.centers}
\title{Retrieve the Model Centers}
\usage{
h2o.centers(object)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}
}
\description{
Retrieve the Model Centers
}
\examples{
\dontrun{
library(h2o)
h2o.init()
fr <- h2o.importFile("http://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv")
h2o.ceiling(fr[, 1])
}
}
