% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.centersSTD}
\alias{h2o.centersSTD}
\title{Retrieve the Model Centers STD}
\usage{
h2o.centersSTD(object)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}
}
\description{
Retrieve the Model Centers STD
}

