% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.cluster_sizes}
\alias{h2o.cluster_sizes}
\title{Retrieve the cluster sizes
If "train", "valid", and "xval" parameters are FALSE (default), then the training cluster sizes value is returned. If more
than one parameter is set to TRUE, then a named list of cluster size vectors are returned, where the names are "train", "valid"
or "xval".}
\usage{
h2o.cluster_sizes(object, train = FALSE, valid = FALSE, xval = FALSE)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}

\item{train}{Retrieve the training cluster sizes}

\item{valid}{Retrieve the validation cluster sizes}

\item{xval}{Retrieve the cross-validation cluster sizes}
}
\description{
Retrieve the cluster sizes
If "train", "valid", and "xval" parameters are FALSE (default), then the training cluster sizes value is returned. If more
than one parameter is set to TRUE, then a named list of cluster size vectors are returned, where the names are "train", "valid"
or "xval".
}

