% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{h2o.clusterIsUp}
\alias{h2o.clusterIsUp}
\title{Determine if an H2O cluster is up or not}
\usage{
h2o.clusterIsUp(conn = h2o.getConnection())
}
\arguments{
\item{conn}{H2OConnection object}
}
\value{
TRUE if the cluster is up; FALSE otherwise
}
\description{
Determine if an H2O cluster is up or not
}
