% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.clusterStatus}
\alias{h2o.clusterStatus}
\title{Return the status of the cluster}
\usage{
h2o.clusterStatus(conn = h2o.getConnection())
}
\arguments{
\item{conn}{the \linkS4class{H2OConnection} object containing the IP address
and port of the server running H2O.}
}
\description{
Retrieve information on the status of the cluster running H2O.
}
\examples{
localH2O <- h2o.init()
h2o.clusterStatus(localH2O)
}
\seealso{
\linkS4class{H2OConnection}, \code{\link{h2o.init}}
}

