% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.coef}
\alias{h2o.coef}
\title{Return the coefficients that can be applied to the non-standardized data.}
\usage{
h2o.coef(object)
}
\arguments{
\item{object}{an \linkS4class{H2OModel} object.}
}
\description{
Note: standardize = True by default. If set to False, then coef() returns the coefficients that are fit directly.
}

