% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.colnames}
\alias{h2o.colnames}
\title{Return column names of an H2OFrame}
\usage{
h2o.colnames(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Return column names of an H2OFrame
}
\seealso{
\code{\link[base]{colnames}} for the base R implementation.
}

