% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.cosh}
\alias{h2o.cosh}
\title{Compute the hyperbolic cosine of x}
\usage{
h2o.cosh(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the hyperbolic cosine of x
}
\seealso{
\code{\link[base]{cosh}} for the base R implementation.
}
