% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.cross_validation_fold_assignment}
\alias{h2o.cross_validation_fold_assignment}
\title{Retrieve the cross-validation fold assignment}
\usage{
h2o.cross_validation_fold_assignment(object)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}
}
\value{
Returns a H2OFrame
}
\description{
Retrieve the cross-validation fold assignment
}
