% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.cross_validation_models}
\alias{h2o.cross_validation_models}
\title{Retrieve the cross-validation models}
\usage{
h2o.cross_validation_models(object)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}
}
\value{
Returns a list of H2OModel objects
}
\description{
Retrieve the cross-validation models
}
