% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.cummin}
\alias{h2o.cummin}
\title{Return the cumulative min over a column or across a row}
\usage{
h2o.cummin(x, axis = 0)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{axis}{An int that indicates whether to do down a column (0) or across a row (1).}
}
\description{
Return the cumulative min over a column or across a row
}
\seealso{
\code{\link[base]{cummin}} for the base R implementation.
}

