% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.cumprod}
\alias{h2o.cumprod}
\title{Cumulative Product}
\usage{
h2o.cumprod(x)
}
\arguments{
\item{x}{An \linkS4class{H2OFrame} object.}
}
\description{
Obtain the cumulative product of a column.
}
\examples{
localH2O <- h2o.init()
fr <- as.h2o(iris)
h2o.cumprod(fr[,1])
}

