% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.cumsum}
\alias{h2o.cumsum}
\title{Cumulative Sum}
\usage{
h2o.cumsum(x)
}
\arguments{
\item{x}{An \linkS4class{H2OFrame} object.}
}
\description{
Obtain the cumulative sum of a column.
}
\examples{
localH2O <- h2o.init()
fr <- as.h2o(iris)
h2o.cumsum(fr[,1])
}

