% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.day}
\alias{day}
\alias{day.H2OFrame}
\alias{h2o.day}
\title{Convert Milliseconds to Day of Month in H2O Datasets}
\usage{
h2o.day(x)

day(x)

\method{day}{H2OFrame}(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\value{
An H2OFrame object containing the entries of \code{x} converted to days of
        the month.
}
\description{
Converts the entries of an H2OFrame object from milliseconds to days of the month
(on a 1 to 31 scale).
}
\seealso{
\code{\link{h2o.month}}
}

