% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.dayOfWeek}
\alias{dayOfWeek}
\alias{dayOfWeek.H2OFrame}
\alias{h2o.dayOfWeek}
\title{Convert Milliseconds to Day of Week in H2O Datasets}
\usage{
h2o.dayOfWeek(x)

dayOfWeek(x)

\method{dayOfWeek}{H2OFrame}(x)
}
\arguments{
\item{x}{An \linkS4class{H2OFrame} object.}
}
\value{
A \linkS4class{H2OFrame} object containing the entries of \code{x} converted to days of
        the week.
}
\description{
Converts the entries of a \linkS4class{H2OFrame} object from milliseconds to days of the week
(on a 0 to 6 scale).
}
\seealso{
\code{\link{h2o.day}, \link{h2o.month}}
}

