% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.difflag1}
\alias{h2o.difflag1}
\title{Conduct a lag 1 transform on a numeric H2OFrame column}
\usage{
h2o.difflag1(object)
}
\arguments{
\item{object}{H2OFrame object}
}
\value{
Returns an H2OFrame object.
}
\description{
Conduct a lag 1 transform on a numeric H2OFrame column
}
