% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.dim}
\alias{h2o.dim}
\title{Returns the number of rows and columns for an H2OFrame object.}
\usage{
h2o.dim(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Returns the number of rows and columns for an H2OFrame object.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/junit/cars_20mpg.csv"
cars <- h2o.importFile(f)
h2o.dim(cars)
}
}
\seealso{
\code{\link[base]{dim}} for the base R implementation.
}
