% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.dim}
\alias{h2o.dim}
\title{Returns the number of rows and columns for an H2OFrame object.}
\usage{
h2o.dim(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Returns the number of rows and columns for an H2OFrame object.
}
\seealso{
\code{\link[base]{dim}} for the base R implementation.
}

