% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.dimnames}
\alias{h2o.dimnames}
\title{Column names of an H2OFrame}
\usage{
h2o.dimnames(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Column names of an H2OFrame
}
\seealso{
\code{\link[base]{dimnames}} for the base R implementation.
}
