% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{h2o.downloadAllLogs}
\alias{h2o.downloadAllLogs}
\title{Download H2O Log Files to Disk}
\usage{
h2o.downloadAllLogs(dirname = ".", filename = NULL)
}
\arguments{
\item{dirname}{(Optional) A character string indicating the directory that the log file should be saved in.}

\item{filename}{(Optional) A character string indicating the name that the log file should be saved to. Note that the saved format is .zip, so the file name must include the .zip extension.}
}
\description{
\code{h2o.downloadAllLogs} downloads all H2O log files to local disk in .zip format. Generally used for debugging purposes.
}
\examples{
\dontrun{
h2o.downloadAllLogs(dirname='./your_directory_name/', filename = 'autoh2o_log.zip')
}
}
