% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.entropy}
\alias{h2o.entropy}
\title{Shannon entropy}
\usage{
h2o.entropy(x)
}
\arguments{
\item{x}{The column on which to calculate the entropy.}
}
\description{
Return the Shannon entropy of a string column. If the string is empty, the entropy is 0.
}

