% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.exp}
\alias{h2o.exp}
\title{Compute the exponential function of x}
\usage{
h2o.exp(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the exponential function of x
}
\seealso{
\code{\link[base]{exp}} for the base R implementation.
}

