% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.exportHDFS}
\alias{h2o.exportHDFS}
\title{Export a Model to HDFS}
\usage{
h2o.exportHDFS(object, path, force = FALSE)
}
\arguments{
\item{object}{an \linkS4class{H2OModel} class object.}

\item{path}{The path to write the model to. Must include the driectory and
filename.}

\item{force}{logical, indicates how to deal with files that already exist.}
}
\description{
Exports an \linkS4class{H2OModel} to HDFS.
}

