% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.find_row_by_threshold}
\alias{h2o.find_row_by_threshold}
\title{Find the threshold, give the max metric. No duplicate thresholds allowed}
\usage{
h2o.find_row_by_threshold(object, threshold)
}
\arguments{
\item{object}{H2OBinomialMetrics}

\item{threshold}{number between 0 and 1}
}
\description{
Find the threshold, give the max metric. No duplicate thresholds allowed
}

