% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.findSynonyms}
\alias{h2o.findSynonyms}
\title{Find synonyms using a word2vec model.}
\usage{
h2o.findSynonyms(word2vec, word, count = 20)
}
\arguments{
\item{word2vec}{A word2vec model.}

\item{word}{A single word to find synonyms for.}

\item{count}{The top `count` synonyms will be returned.}
}
\description{
Find synonyms using a word2vec model.
}

