% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.floor}
\alias{h2o.floor}
\title{floor takes a single numeric argument x and returns a numeric
vector containing the largest integers not greater than the
corresponding elements of x.}
\usage{
h2o.floor(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
floor takes a single numeric argument x and returns a numeric
vector containing the largest integers not greater than the
corresponding elements of x.
}
\seealso{
\code{\link[base]{floor}} for the base R implementation.
}

