% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.floor}
\alias{h2o.floor}
\title{Take a single numeric argument and return a numeric vector with the largest integers}
\usage{
h2o.floor(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
floor takes a single numeric argument x and returns a numeric
vector containing the largest integers not greater than the
corresponding elements of x.
}
\seealso{
\code{\link[base]{floor}} for the base R implementation.
}

