% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.getFrame}
\alias{h2o.getFrame}
\title{Get an R Reference to an H2O Dataset, that will NOT be GC'd by default}
\usage{
h2o.getFrame(id)
}
\arguments{
\item{id}{A string indicating the unique frame of the dataset to retrieve.}
}
\description{
Get the reference to a frame with the given id in the H2O instance.
}

