% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm.R
\name{h2o.getGLMFullRegularizationPath}
\alias{h2o.getGLMFullRegularizationPath}
\title{Extract full regularization path from glm model (assuming it was run with lambda search option)}
\usage{
h2o.getGLMFullRegularizationPath(model)
}
\arguments{
\item{model}{an \linkS4class{H2OModel} corresponding from a \code{h2o.glm} call.}
}
\description{
Extract full regularization path from glm model (assuming it was run with lambda search option)
}

