% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.getId}
\alias{h2o.getId}
\title{Get back-end distributed key/value store id from an H2OFrame.}
\usage{
h2o.getId(x)
}
\arguments{
\item{x}{An H2OFrame}
}
\value{
The id of the H2OFrame
}
\description{
Get back-end distributed key/value store id from an H2OFrame.
}

