% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.getId}
\alias{h2o.getId}
\title{Get back-end distributed key/value store id from an H2OFrame.}
\usage{
h2o.getId(x)
}
\arguments{
\item{x}{An H2OFrame}
}
\value{
The id of the H2OFrame
}
\description{
Get back-end distributed key/value store id from an H2OFrame.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.getId(iris)
}
}
