% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvstore.R
\name{h2o.getModel}
\alias{h2o.getModel}
\title{Get an R reference to an H2O model}
\usage{
h2o.getModel(model_id)
}
\arguments{
\item{model_id}{A string indicating the unique model_id of the model to retrieve.}
}
\value{
Returns an object that is a subclass of \linkS4class{H2OModel}.
}
\description{
Returns a reference to an existing model in the H2O instance.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

iris_hf <- as.h2o(iris)
model_id <- h2o.gbm(x = 1:4, y = 5, training_frame = iris_hf)@model_id
model_retrieved <- h2o.getModel(model_id)
}
}
