% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.getTypes}
\alias{h2o.getTypes}
\title{Get the types-per-column}
\usage{
h2o.getTypes(x)
}
\arguments{
\item{x}{An H2OFrame}
}
\value{
A list of types per column
}
\description{
Get the types-per-column
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "http://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.getTypes(iris)
}
}
