% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.hist}
\alias{h2o.hist}
\title{Compute A Histgram}
\usage{
h2o.hist(x, breaks = "Sturges", plot = TRUE)
}
\arguments{
\item{x}{A single numeric column from an H2OFrame.}

\item{breaks}{Can be one of the following:
A string: "Sturges", "Rice", "sqrt", "Doane", "FD", "Scott"
A single number for the number of breaks splitting the range of the vec into number of breaks bins of equal width
A vector of numbers giving the split points, e.g., c(-50,213.2123,9324834)}

\item{plot}{A logical value indicating whether or not a plot should be generated (default is TRUE).}
}
\description{
Compute a histogram over a numeric column. If breaks=="FD", the MAD is used over the IQR
in computing bin width.
}

