% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.hit_ratio_table}
\alias{h2o.hit_ratio_table}
\title{Retrieve the Hit Ratios
If "train", "valid", and "xval" parameters are FALSE (default), then the training Hit Ratios value is returned. If more
than one parameter is set to TRUE, then a named list of Hit Ratio tables are returned, where the names are "train", "valid"
or "xval".}
\usage{
h2o.hit_ratio_table(object, train = FALSE, valid = FALSE, xval = FALSE,
  ...)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}

\item{train}{Retrieve the training Hit Ratio}

\item{valid}{Retrieve the validation Hit Ratio}

\item{xval}{Retrieve the cross-validation Hit Ratio}

\item{\dots}{further arguments to be passed on (currently unimplemented)}
}
\description{
Retrieve the Hit Ratios
If "train", "valid", and "xval" parameters are FALSE (default), then the training Hit Ratios value is returned. If more
than one parameter is set to TRUE, then a named list of Hit Ratio tables are returned, where the names are "train", "valid"
or "xval".
}

