% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.hit_ratio_table}
\alias{h2o.hit_ratio_table}
\title{Retrieve the Hit Ratios}
\usage{
h2o.hit_ratio_table(object, train = FALSE, valid = FALSE, xval = FALSE)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}

\item{train}{Retrieve the training Hit Ratio}

\item{valid}{Retrieve the validation Hit Ratio}

\item{xval}{Retrieve the cross-validation Hit Ratio}
}
\description{
If "train", "valid", and "xval" parameters are FALSE (default), then the training Hit Ratios value is returned. If more
than one parameter is set to TRUE, then a named list of Hit Ratio tables are returned, where the names are "train", "valid"
or "xval".
}
