% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.ischaracter}
\alias{h2o.ischaracter}
\title{Check if character}
\usage{
h2o.ischaracter(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Check if character
}
\seealso{
\code{\link[base]{is.character}} for the base R implementation.
}

