% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.isnumeric}
\alias{h2o.isnumeric}
\title{Check if numeric}
\usage{
h2o.isnumeric(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Check if numeric
}
\seealso{
\code{\link[base]{is.numeric}} for the base R implementation.
}
