% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.listTimezones}
\alias{h2o.listTimezones}
\title{List all of the Time Zones Acceptable by the H2O Cloud.}
\usage{
h2o.listTimezones(conn = h2o.getConnection())
}
\arguments{
\item{conn}{An H2OConnection object.}
}
\description{
List all of the Time Zones Acceptable by the H2O Cloud.
}

