% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.log10}
\alias{h2o.log10}
\title{Compute the log10 of x}
\usage{
h2o.log10(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the log10 of x
}
\seealso{
\code{\link[base]{log10}} for the base R implementation.
}

