% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.logloss}
\alias{h2o.logloss}
\title{Retrieve the Log Loss Value}
\usage{
h2o.logloss(object, train = FALSE, valid = FALSE, xval = FALSE)
}
\arguments{
\item{object}{a \linkS4class{H2OModelMetrics} object of the correct type.}

\item{train}{Retrieve the training Log Loss}

\item{valid}{Retrieve the validation Log Loss}

\item{xval}{Retrieve the cross-validation Log Loss}
}
\description{
Retrieves the log loss output for a \linkS4class{H2OBinomialMetrics} or
\linkS4class{H2OMultinomialMetrics} object
If "train", "valid", and "xval" parameters are FALSE (default), then the training Log Loss value is returned. If more
than one parameter is set to TRUE, then a named vector of Log Losses are returned, where the names are "train", "valid"
or "xval".
}
