% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glm.R
\name{h2o.makeGLMModel}
\alias{h2o.makeGLMModel}
\title{Set betas of an existing H2O GLM Model}
\usage{
h2o.makeGLMModel(model, beta)
}
\arguments{
\item{model}{an \linkS4class{H2OModel} corresponding from a \code{h2o.glm} call.}

\item{beta}{a new set of betas (a named vector)}
}
\description{
This function allows setting betas of an existing glm model.
}

