% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.median}
\alias{h2o.median}
\alias{median.Frame}
\title{H2O Median}
\usage{
h2o.median(x, na.rm = TRUE)

\method{median}{Frame}(x, na.rm = TRUE)
}
\arguments{
\item{x}{An H2O Frame object.}

\item{na.rm}{a logical, indicating whether na's are omitted.}
}
\description{
Compute the median of a Frame.
}
\examples{
\donttest{
h2o.init()
prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath, destination_frame = "prostate.hex")
}
}

