% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.merge}
\alias{h2o.merge}
\title{Merge Two H2O Data Frames}
\usage{
h2o.merge(x, y, all.x = FALSE, all.y = FALSE)
}
\arguments{
\item{x,y}{\linkS4class{H2OFrame} objects}

\item{all.x}{a logical value indicating whether or not shared values are
preserved or ignored in \code{x}.}

\item{all.y}{a logical value indicating whether or not shared values are
preserved or ignored in \code{y}.}
}
\description{
Merges two \linkS4class{H2OFrame} objects by shared column names. Unlike the
base R implementation, \code{h2o.merge} only supports merging through shared
column names.
}
\details{
In order for \code{h2o.merge} to work in multinode clusters, one of the
datasets must be small enough to exist in every node. Currently, this
function only supports \code{all.x = TRUE}. All other permutations will fail.
}
\examples{
h2o.init()
left <- data.frame(fruit = c('apple', 'orange', 'banana', 'lemon', 'strawberry', 'blueberry'),
color = c('red', 'orange', 'yellow', 'yellow', 'red', 'blue'))
right <- data.frame(fruit = c('apple', 'orange', 'banana', 'lemon', 'strawberry', 'watermelon'),
citrus = c(FALSE, TRUE, FALSE, TRUE, FALSE, FALSE))
l.hex <- as.h2o(left)
r.hex <- as.h2o(right)
left.hex <- h2o.merge(l.hex, r.hex, all.x = TRUE)
}

