% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.merge}
\alias{h2o.merge}
\title{Merge Two H2O Data Frames}
\usage{
h2o.merge(x, y, by = intersect(names(x), names(y)), by.x = by, by.y = by,
  all = FALSE, all.x = all, all.y = all, method = "hash")
}
\arguments{
\item{x,y}{H2OFrame objects}

\item{by}{columns used for merging by default the common names}

\item{by.x}{x columns used for merging by name or number}

\item{by.y}{y columns used for merging by name or number}

\item{all}{TRUE includes all rows in x and all rows in y even if there is no match to the other}

\item{all.x}{If all.x is true, all rows in the x will be included, even if there is no matching
row in y, and vice-versa for all.y.}

\item{all.y}{see all.x}

\item{method}{auto, radix, or hash (default)}
}
\description{
Merges two H2OFrame objects with the same arguments and meanings
as merge() in base R.
}
\examples{
\donttest{
h2o.init()
left <- data.frame(fruit = c('apple', 'orange', 'banana', 'lemon', 'strawberry', 'blueberry'),
color <- c('red', 'orange', 'yellow', 'yellow', 'red', 'blue'))
right <- data.frame(fruit = c('apple', 'orange', 'banana', 'lemon', 'strawberry', 'watermelon'),
citrus <- c(FALSE, TRUE, FALSE, TRUE, FALSE, FALSE))
l.hex <- as.h2o(left)
r.hex <- as.h2o(right)
left.hex <- h2o.merge(l.hex, r.hex, all.x = TRUE)
}
}

