% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.min}
\alias{h2o.min}
\title{Returns the minima of the input values.}
\usage{
h2o.min(x, na.rm = FALSE)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{\code{logical}. indicating whether missing values should be removed.}
}
\description{
Returns the minima of the input values.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.min(iris["sepal_len"], na.rm = TRUE)
}
}
\seealso{
\code{\link[base]{Extremes}} for the base R implementation, \code{min()}.
}
