% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.month}
\alias{h2o.month}
\alias{month}
\alias{month.H2OFrame}
\title{Convert Milliseconds to Months in H2O Datasets}
\usage{
h2o.month(x)

month(x)

\method{month}{H2OFrame}(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\value{
An H2OFrame object containing the entries of \code{x} converted to months of
        the year.
}
\description{
Converts the entries of an H2OFrame object from milliseconds to months (on a 1 to
12 scale).
}
\seealso{
\code{\link{h2o.year}}
}
