% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.na_omit}
\alias{h2o.na_omit}
\title{Remove Rows With NAs}
\usage{
h2o.na_omit(object, ...)
}
\arguments{
\item{object}{H2OFrame object}

\item{...}{Ignored}
}
\value{
Returns an H2OFrame object containing non-NA rows.
}
\description{
Remove Rows With NAs
}
\examples{
\dontrun{
library(h2o)
h2o.init()

frame <- h2o.createFrame(rows = 6, cols = 2,
                         categorical_fraction = 0.0, 
                         missing_fraction = 0.7, 
                         seed = 123)
h2o.na_omit(frame)
}
}
