% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.nacnt}
\alias{h2o.nacnt}
\title{Count of NAs per column}
\usage{
h2o.nacnt(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Gives the count of NAs per column.
}
\examples{
\donttest{
h2o.init()
iris.hex <- as.h2o(iris)
h2o.nacnt(iris.hex)  # should return all 0s
h2o.insertMissingValues(iris.hex)
h2o.nacnt(iris.hex)
}
}

