% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.names}
\alias{h2o.names}
\title{Column names of an H2OFrame}
\usage{
h2o.names(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Column names of an H2OFrame
}
\seealso{
\code{\link[base]{names}} for the base R implementation.
}

