% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.ncol}
\alias{h2o.ncol}
\title{Return the number of columns present in x.}
\usage{
h2o.ncol(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Return the number of columns present in x.
}
\seealso{
\code{\link[base]{ncol}} for the base R implementation.
}

