% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.ncol}
\alias{h2o.ncol}
\title{Return the number of columns present in x.}
\usage{
h2o.ncol(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Return the number of columns present in x.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "http://h2o-public-test-data.s3.amazonaws.com/smalldata/iris/iris_train.csv"
iris <- h2o.importFile(f)
h2o.ncol(iris)
}
}
\seealso{
\code{\link[base]{ncol}} for the base R implementation.
}
