% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connection.R
\name{h2o.networkTest}
\alias{h2o.networkTest}
\title{View Network Traffic Speed}
\usage{
h2o.networkTest()
}
\value{
Returns a table listing the network speed for 1B, 10KB, and 10MB.
}
\description{
View speed with various file sizes.
}

