% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.nlevels}
\alias{h2o.nlevels}
\title{Get the number of factor levels for this frame.}
\usage{
h2o.nlevels(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Get the number of factor levels for this frame.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/junit/cars_20mpg.csv"
cars <- h2o.importFile(f)
h2o.nlevels(cars)
}
}
\seealso{
\code{\link[base]{nlevels}} for the base R method.
}
