% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.null_deviance}
\alias{h2o.null_deviance}
\title{Retrieve the null deviance
If "train", "valid", and "xval" parameters are FALSE (default), then the training null deviance value is returned. If more
than one parameter is set to TRUE, then a named vector of null deviances are returned, where the names are "train", "valid"
or "xval".}
\usage{
h2o.null_deviance(object, train = FALSE, valid = FALSE, xval = FALSE)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}}

\item{train}{Retrieve the training null deviance}

\item{valid}{Retrieve the validation null deviance}

\item{xval}{Retrieve the cross-validation null deviance}
}
\description{
Retrieve the null deviance
If "train", "valid", and "xval" parameters are FALSE (default), then the training null deviance value is returned. If more
than one parameter is set to TRUE, then a named vector of null deviances are returned, where the names are "train", "valid"
or "xval".
}

