% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.null_dof}
\alias{h2o.null_dof}
\title{Retrieve the null degrees of freedom}
\usage{
h2o.null_dof(object, train = FALSE, valid = FALSE, xval = FALSE)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}}

\item{train}{Retrieve the training null degrees of freedom}

\item{valid}{Retrieve the validation null degrees of freedom}

\item{xval}{Retrieve the cross-validation null degrees of freedom}
}
\description{
If "train", "valid", and "xval" parameters are FALSE (default), then the training null degrees of freedom value is returned. If more
than one parameter is set to TRUE, then a named vector of null degrees of freedom are returned, where the names are "train", "valid"
or "xval".
}
\examples{
\dontrun{
library(h2o)
h2o.init()

prostate_path <- system.file("extdata", "prostate.csv", package = "h2o")
prostate <- h2o.importFile(prostate_path)
prostate[,2] <- as.factor(prostate[,2])
prostate_glm <- h2o.glm(y = "CAPSULE", x = c("AGE","RACE","PSA","DCAPS"), 
                        training_frame = prostate, family = "binomial", nfolds = 0, 
                        alpha = 0.5, lambda_search = FALSE)
h2o.null_dof(prostate_glm, train = TRUE)
}
}
