% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.null_dof}
\alias{h2o.null_dof}
\title{Retrieve the null degrees of freedom
If "train", "valid", and "xval" parameters are FALSE (default), then the training null degrees of freedom value is returned. If more
than one parameter is set to TRUE, then a named vector of null degrees of freedom are returned, where the names are "train", "valid"
or "xval".}
\usage{
h2o.null_dof(object, train = FALSE, valid = FALSE, xval = FALSE, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}}

\item{train}{Retrieve the training null degrees of freedom}

\item{valid}{Retrieve the validation null degrees of freedom}

\item{xval}{Retrieve the cross-validation null degrees of freedom}

\item{\dots}{further arguments to be passed to/from this method.}
}
\description{
Retrieve the null degrees of freedom
If "train", "valid", and "xval" parameters are FALSE (default), then the training null degrees of freedom value is returned. If more
than one parameter is set to TRUE, then a named vector of null degrees of freedom are returned, where the names are "train", "valid"
or "xval".
}

