% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.null_dof}
\alias{h2o.null_dof}
\title{Retrieve the null degrees of freedom}
\usage{
h2o.null_dof(object, valid = FALSE, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}}

\item{valid}{Retrieve the validation metric.}

\item{\dots}{further arguments to be passed to/from this method.}
}
\description{
Retrieve the null degrees of freedom
}

