% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.parseSetup}
\alias{h2o.parseSetup}
\title{Get a parse setup back for the staged data.}
\usage{
h2o.parseSetup(data, destination_frame = "", header = NA, sep = "",
  col.names = NULL, col.types = NULL, na.strings = NULL)
}
\arguments{
\item{data}{An \linkS4class{H2ORawData} object to be parsed.}

\item{destination_frame}{(Optional) The hex key assigned to the parsed file.}

\item{header}{(Optional) A logical value indicating whether the first row is
the column header. If missing, H2O will automatically try to detect
the presence of a header.}

\item{sep}{(Optional) The field separator character. Values on each line of
the file are separated by this character. If \code{sep = ""}, the
parser will automatically detect the separator.}

\item{col.names}{(Optional) A \linkS4class{H2OFrame} object containing a
single delimited line with the column names for the file.}

\item{col.types}{(Optional) A vector specifying the types to attempt to force
over columns.}

\item{na.strings}{(Optional) H2O will interpret these strings as missing.}
}
\description{
Get a parse setup back for the staged data.
}

