% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.prod}
\alias{h2o.prod}
\title{Return the product of all the values present in its arguments.}
\usage{
h2o.prod(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Return the product of all the values present in its arguments.
}
\seealso{
\code{\link[base]{prod}} for the base R implementation.
}

