% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.relevel}
\alias{h2o.relevel}
\title{Reorders levels of an H2O factor, similarly to standard R's relevel.}
\usage{
h2o.relevel(x, y)
}
\arguments{
\item{x}{factor column in h2o frame}

\item{y}{reference level (string)}
}
\value{
new reordered factor column
}
\description{
The levels of a factor are reordered os that the reference level is at level 0, remaining levels are moved down as needed.
}
\examples{
\dontrun{
library(h2o)
h2o.init()

# Convert iris dataset to an H2OFrame
iris_hf <- as.h2o(iris)
# Look at current ordering of the Species column levels
h2o.levels(iris_hf["Species"])
# "setosa"     "versicolor" "virginica" 
# Change the reference level to "virginica"
iris_hf["Species"] <- h2o.relevel(x = iris_hf["Species"], y = "virginica")
# Observe new ordering
h2o.levels(iris_hf["Species"])
# "virginica"  "setosa"     "versicolor"
}
}
