% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.removeVecs}
\alias{h2o.removeVecs}
\title{Delete Columns from an H2OFrame}
\usage{
h2o.removeVecs(data, cols)
}
\arguments{
\item{data}{The H2OFrame.}

\item{cols}{The columns to remove.}
}
\description{
Delete the specified columns from the H2OFrame.  Returns an H2OFrame without the specified
columns.
}
