% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.removeVecs}
\alias{h2o.removeVecs}
\title{Delete Columns from a H2OFrame}
\usage{
h2o.removeVecs(data, cols)
}
\arguments{
\item{data}{The H2OFrame.}

\item{cols}{The columns to remove.}
}
\description{
Delete the specified columns from the H2OFrame. Returns a H2OFrame without the specified
columns. This will trigger any lazy computation of the frame, and has side-effects.
}

