% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.rep_len}
\alias{h2o.rep_len}
\title{Replicate Elements of Vectors or Lists into H2O}
\usage{
h2o.rep_len(x, length.out)
}
\arguments{
\item{x}{a vector (of any mode including a list) or a factor}

\item{length.out}{non negative integer. The desired length of the output
vector.}
}
\value{
Creates a Frame vector of the same type as x
}
\description{
\code{h2o.rep} performs just as \code{rep} does. It replicates the values in
\code{x} in the H2O backend.
}

