% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.residual_deviance}
\alias{h2o.residual_deviance}
\title{Retrieve the residual deviance}
\usage{
h2o.residual_deviance(object, valid = FALSE, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}}

\item{valid}{Retrieve the validation metric.}

\item{\dots}{further arguments to be passed to/from this method.}
}
\description{
Retrieve the residual deviance
}

