% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaglm.R
\name{h2o.result}
\alias{h2o.result}
\title{Extract and return ANOVA Table as an H2OFrame}
\usage{
h2o.result(model)
}
\arguments{
\item{model}{an H2OANOVAGLM.}
}
\description{
Extract and return ANOVA Table as an H2OFrame
}
