% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.rm}
\alias{h2o.rm}
\title{Delete Objects In H2O}
\usage{
h2o.rm(ids, conn = h2o.getConnection())
}
\arguments{
\item{ids}{The hex key associated with the object to be removed.}

\item{conn}{An \linkS4class{H2OConnection} object containing the IP address and port number of the H2O server.}
}
\description{
Remove the h2o Big Data object(s) having the key name(s) from ids.
}
\seealso{
\code{\link{h2o.assign}}, \code{\link{h2o.ls}}
}

