% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.sdev}
\alias{h2o.sdev}
\title{Retrieve the standard deviations of principal components}
\usage{
h2o.sdev(object)
}
\arguments{
\item{object}{An \linkS4class{H2ODimReductionModel} object.}
}
\description{
Retrieve the standard deviations of principal components
}

