% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.setTimezone}
\alias{h2o.setTimezone}
\title{Set the Time Zone on the H2O cluster}
\usage{
h2o.setTimezone(tz)
}
\arguments{
\item{tz}{The desired timezone.}
}
\description{
Set the Time Zone on the H2O cluster
}
