% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.signif}
\alias{h2o.signif}
\alias{signif}
\title{Round doubles/floats to the given number of significant digits.}
\usage{
h2o.signif(x, digits = 6)

signif(x, digits = 6)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{digits}{Number of significant digits to round doubles/floats.}
}
\description{
Round doubles/floats to the given number of significant digits.
}
\seealso{
\code{\link[base]{signif}} for the base R implementation.
}
