% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.skewness}
\alias{h2o.skewness}
\alias{skewness.H2OFrame}
\title{Skewness of a column}
\usage{
h2o.skewness(x, ..., na.rm = TRUE)

skewness.H2OFrame(x, ..., na.rm = TRUE)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{...}{Further arguments to be passed from or to other methods.}

\item{na.rm}{A logical value indicating whether \code{NA} or missing values should be stripped before the computation.}
}
\value{
Returns a list containing the skewness for each column (NaN for non-numeric columns).
}
\description{
Obtain the skewness of a column of a parsed H2O data object.
}
\examples{
\donttest{
h2o.init()
prosPath <- system.file("extdata", "prostate.csv", package="h2o")
prostate.hex <- h2o.uploadFile(path = prosPath)
h2o.skewness(prostate.hex$AGE)
}
}
