% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.sub}
\alias{h2o.sub}
\title{String Substitute}
\usage{
h2o.sub(pattern, replacement, x, ignore.case = FALSE)
}
\arguments{
\item{pattern}{The pattern to replace.}

\item{replacement}{The replacement pattern.}

\item{x}{The column on which to operate.}

\item{ignore.case}{Case sensitive or not}
}
\description{
Creates a copy of the target column in which each string has the first occurence of
the regex pattern replaced with the replacement substring.
}
\examples{
\donttest{
library(h2o)
h2o.init()
string_to_sub <- as.h2o("r tutorial")
sub_string <- h2o.sub("r ","H2O ",string_to_sub)
}
}

