% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.sum}
\alias{h2o.sum}
\title{Compute the frame's sum by-column (or by-row).}
\usage{
h2o.sum(x, na.rm = FALSE, axis = 0, return_frame = FALSE)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{\code{logical}. indicating whether missing values should be removed.}

\item{axis}{An int that indicates whether to do down a column (0) or across a row (1). For row or column sums, the \code{return_frame} parameter must be TRUE.}

\item{return_frame}{A boolean that indicates whether to return an H2O frame or one single aggregated value. Default is FALSE.}
}
\description{
Compute the frame's sum by-column (or by-row).
}
\examples{
\dontrun{
library(h2o)
h2o.init()

frame <- h2o.createFrame(rows = 6, cols = 2,
                         categorical_fraction = 0.0, 
                         missing_fraction = 0.7, 
                         seed = 123)
h2o.sum(frame["C1"], na.rm = TRUE, axis = 0, return_frame = TRUE)
}
}
\seealso{
\code{\link[base]{sum}} for the base R implementation.
}
