% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.sum}
\alias{h2o.sum}
\title{Compute the frame's sum by-column (or by-row).}
\usage{
h2o.sum(x, na.rm = FALSE, axis = 0, return_frame = FALSE)
}
\arguments{
\item{x}{An H2OFrame object.}

\item{na.rm}{\code{logical}. indicating whether missing values should be removed.}

\item{axis}{An int that indicates whether to do down a column (0) or across a row (1).}

\item{return_frame}{A boolean that indicates whether to return an H2O frame or a list. Default is FALSE.}
}
\description{
Compute the frame's sum by-column (or by-row).
}
\seealso{
\code{\link[base]{sum}} for the base R implementation.
}

