% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.tanh}
\alias{h2o.tanh}
\title{Compute the hyperbolic tangent of x}
\usage{
h2o.tanh(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
Compute the hyperbolic tangent of x
}
\examples{
\dontrun{
library(h2o)
h2o.init()

frame <- h2o.createFrame(rows = 6, cols = 2,
                         categorical_fraction = 0.0, 
                         missing_fraction = 0.7, 
                         seed = 123)
h2o.tanh(frame)
}
}
\seealso{
\code{\link[base]{Hyperbolic}} for the base R implementation, \code{tanh()}.
}
