% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.topN}
\alias{h2o.topN}
\title{H2O topN}
\usage{
h2o.topN(x, column, nPercent)
}
\arguments{
\item{x}{an H2OFrame}

\item{column}{is a column name or column index to grab the top N percent value from}

\item{nPercent}{is a top percentage value to grab}
}
\value{
An H2OFrame with 2 columns.  The first column is the original row indices, second column contains the topN values
}
\description{
Extract the top N percent  of values of a column and return it in a H2OFrame.
}

