% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.tot_withinss}
\alias{h2o.tot_withinss}
\title{Get the total within cluster sum of squares.
If "train", "valid", and "xval" parameters are FALSE (default), then the training tot_withinss value is returned. If more
than one parameter is set to TRUE, then a named vector of tot_withinss' are returned, where the names are "train", "valid"
or "xval".}
\usage{
h2o.tot_withinss(object, train = FALSE, valid = FALSE, xval = FALSE, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}

\item{train}{Retrieve the training total within cluster sum of squares}

\item{valid}{Retrieve the validation total within cluster sum of squares}

\item{xval}{Retrieve the cross-validation total within cluster sum of squares}

\item{\dots}{further arguments to be passed on (currently unimplemented)}
}
\description{
Get the total within cluster sum of squares.
If "train", "valid", and "xval" parameters are FALSE (default), then the training tot_withinss value is returned. If more
than one parameter is set to TRUE, then a named vector of tot_withinss' are returned, where the names are "train", "valid"
or "xval".
}

