% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.totss}
\alias{h2o.totss}
\title{Get the total sum of squares.
If "train", "valid", and "xval" parameters are FALSE (default), then the training totss value is returned. If more
than one parameter is set to TRUE, then a named vector of totss' are returned, where the names are "train", "valid"
or "xval".}
\usage{
h2o.totss(object, train = FALSE, valid = FALSE, xval = FALSE)
}
\arguments{
\item{object}{An \linkS4class{H2OClusteringModel} object.}

\item{train}{Retrieve the training total sum of squares}

\item{valid}{Retrieve the validation total sum of squares}

\item{xval}{Retrieve the cross-validation total sum of squares}
}
\description{
Get the total sum of squares.
If "train", "valid", and "xval" parameters are FALSE (default), then the training totss value is returned. If more
than one parameter is set to TRUE, then a named vector of totss' are returned, where the names are "train", "valid"
or "xval".
}

