% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.transform}
\alias{h2o.transform}
\title{Transform words to vectors using a word2vec model.}
\usage{
h2o.transform(word2vec, words)
}
\arguments{
\item{word2vec}{A word2vec model.}

\item{words}{An H2OFrame made of a single column containing source words.}
}
\description{
Transform words to vectors using a word2vec model.
}

