% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.trim}
\alias{h2o.trim}
\title{Trim Space}
\usage{
h2o.trim(x)
}
\arguments{
\item{x}{The column whose strings should be trimmed.}
}
\description{
Trim Space
}
\examples{
\donttest{
library(h2o)
h2o.init()
string_to_trim <- as.h2o("r tutorial")
trim_string <- h2o.trim(string_to_trim)
}
}

