% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.trunc}
\alias{h2o.trunc}
\title{trunc takes a single numeric argument x and returns a numeric vector containing the integers
formed by truncating the values in x toward 0.}
\usage{
h2o.trunc(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\description{
trunc takes a single numeric argument x and returns a numeric vector containing the integers
formed by truncating the values in x toward 0.
}
\seealso{
\code{\link[base]{trunc}} for the base R implementation.
}

