% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.varimp}
\alias{h2o.varimp}
\title{Retrieve the variable importance.}
\usage{
h2o.varimp(object, ...)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}

\item{\dots}{further arguments to be passed on (currently unimplemented)}
}
\description{
Retrieve the variable importance.
}

