% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.varimp}
\alias{h2o.varimp}
\title{Retrieve the variable importance.}
\usage{
h2o.varimp(object)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} object.}
}
\description{
Retrieve the variable importance.
}
