% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.varsplits}
\alias{h2o.varsplits}
\title{Retrieve per-variable split information for a given Isolation Forest model.}
\usage{
h2o.varsplits(object)
}
\arguments{
\item{object}{An Isolation Forest model represented by \linkS4class{H2OModel} object.}
}
\description{
Retrieve per-variable split information for a given Isolation Forest model.
}
