% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.week}
\alias{h2o.week}
\alias{week}
\alias{week.Frame}
\title{Convert Milliseconds to Week of Week Year in H2O Datasets}
\usage{
h2o.week(x)

week(x)

\method{week}{Frame}(x)
}
\arguments{
\item{x}{An H2O Frame object.}
}
\value{
A Frame object containing the entries of \code{x} converted to weeks of
        the week year.
}
\description{
Converts the entries of a Frame object from milliseconds to weeks of the week
year (starting from 1).
}
\seealso{
\code{\link{h2o.month}}
}

