% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.week}
\alias{h2o.week}
\alias{week}
\alias{week.H2OFrame}
\title{Convert Milliseconds to Week of Week Year in H2O Datasets}
\usage{
h2o.week(x)

week(x)

\method{week}{H2OFrame}(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\value{
An H2OFrame object containing the entries of \code{x} converted to weeks of
        the week year.
}
\description{
Converts the entries of an H2OFrame object from milliseconds to weeks of the week
year (starting from 1).
}
\seealso{
\code{\link{h2o.month}}
}
