% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{h2o.weights}
\alias{h2o.weights}
\title{Retrieve the respective weight matrix}
\usage{
h2o.weights(object, matrix_id = 1)
}
\arguments{
\item{object}{An \linkS4class{H2OModel} or \linkS4class{H2OModelMetrics}}

\item{matrix_id}{An integer, ranging from 1 to number of layers + 1, that specifies the weight matrix to return.}
}
\description{
Retrieve the respective weight matrix
}
\examples{
\dontrun{
library(h2o)
h2o.init()

f <- "https://s3.amazonaws.com/h2o-public-test-data/smalldata/chicago/chicagoCensus.csv"
census <- h2o.importFile(f)
census[, 1] <- as.factor(census[, 1])
dl_model <- h2o.deeplearning(x = c(1:3), y = 4, training_frame = census,
                            hidden = c(17, 191), 
                            epochs = 1,
                            balance_classes = FALSE,
                            export_weights_and_biases = TRUE)
h2o.weights(dl_model, matrix_id = 1)
}
}
