% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.which}
\alias{h2o.which}
\title{H2O Which}
\usage{
h2o.which(x)
}
\arguments{
\item{x}{An \linkS4class{H2OFrame} object.}
}
\description{
1-based indices similar to R's which
}
\examples{
localH2O <- h2o.init()
fr <- as.h2o(iris)
h2o.which(fr[,5] == "setosa")
}

