% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frame.R
\name{h2o.which}
\alias{h2o.which}
\title{Which indices are TRUE?}
\usage{
h2o.which(x)
}
\arguments{
\item{x}{An H2OFrame object.}
}
\value{
Returns an H2OFrame object.
}
\description{
Give the TRUE indices of a logical object, allowing for array indices.
}
\examples{
\donttest{
h2o.init()
iris.hex <- as.h2o(iris)
h2o.which(iris.hex[,1]==4.4)
}
}
\seealso{
\code{\link[base]{which}} for the base R method.
}

