% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.xgboost.available}
\alias{h2o.xgboost.available}
\title{Ask the H2O server whether a XGBoost model can be built (depends on availability of native backend)
Returns True if a XGBoost model can be built, or False otherwise.}
\usage{
h2o.xgboost.available()
}
\description{
Ask the H2O server whether a XGBoost model can be built (depends on availability of native backend)
Returns True if a XGBoost model can be built, or False otherwise.
}

