% Generated by roxygen2 (4.0.2): do not edit by hand
\name{h2o.year}
\alias{h2o.year}
\alias{year}
\alias{year.H2OFrame}
\title{Convert Milliseconds to Years in H2O Datasets}
\usage{
h2o.year(x)

year(x)

\method{year}{H2OFrame}(x)
}
\arguments{
\item{x}{An \linkS4class{H2OFrame} object.}
}
\value{
A \linkS4class{H2OFrame} object containig the entries of \code{x} converted to years
        starting from 1900, e.g. 69 corresponds to the year 1969.
}
\description{
Convert the entries of a \linkS4class{H2OFrame} object from milliseconds to years, indexed
starting from 1900.
}
\details{
This method calls the function of the MutableDateTime class in Java.
}
\seealso{
\code{\link{h2o.month}}
}

