% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{H2OCoxPHModelSummary-class}
\alias{H2OCoxPHModelSummary-class}
\alias{H2OCoxPHModelSummary}
\alias{show,H2OCoxPHModelSummary-method}
\alias{coef.H2OCoxPHModelSummary}
\title{The H2OCoxPHModelSummary object.}
\usage{
\S4method{show}{H2OCoxPHModelSummary}(object)

\method{coef}{H2OCoxPHModelSummary}(object, ...)
}
\arguments{
\item{object}{An \code{H2OCoxPHModelSummary} object.}

\item{...}{additional arguments to pass on.}
}
\description{
Wrapper object for summary information compatible with survival package.
}
\section{Slots}{

\describe{
\item{\code{summary}}{A \code{list} containing the a summary compatible with CoxPH summary used in the survival package.}
}}

