% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{H2OFrame-Extract}
\alias{H2OFrame-Extract}
\alias{[.H2OFrame}
\alias{[,H2OFrame-method}
\alias{$.H2OFrame}
\alias{[[.H2OFrame}
\alias{$.H2OFrame}
\alias{[[.H2OFrame}
\alias{[<-.H2OFrame}
\alias{$<-.H2OFrame}
\alias{[[<-.H2OFrame}
\title{Extract or Replace Parts of an H2OFrame Object}
\usage{
\method{[}{H2OFrame}(data, row, col, drop = TRUE)

\method{$}{H2OFrame}(x, name)

\method{[[}{H2OFrame}(x, i, exact = TRUE)

\method{$}{H2OFrame}(x, name)

\method{[[}{H2OFrame}(x, i, exact = TRUE)

\method{[}{H2OFrame}(data, row, col, ...) <- value

\method{$}{H2OFrame}(data, name) <- value

\method{[[}{H2OFrame}(data, name) <- value
}
\arguments{
\item{data}{object from which to extract element(s) or in which to replace element(s).}

\item{row}{index specifying row element(s) to extract or replace. Indices are numeric or
character vectors or empty (missing) or will be matched to the names.}

\item{col}{index specifying column element(s) to extract or replace.}

\item{drop}{Unused}

\item{x}{An H2OFrame}

\item{name}{a literal character string or a name (possibly backtick quoted).}

\item{i}{index}

\item{exact}{controls possible partial matching of \code{[[} when extracting
a character}

\item{...}{Further arguments passed to or from other methods.}

\item{value}{To be assigned}
}
\description{
Operators to extract or replace parts of H2OFrame objects.
}
