% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{H2OModelFuture-class}
\alias{H2OModelFuture-class}
\title{H2O Future Model}
\description{
A class to contain the information for background model jobs.
}
\section{Slots}{

\describe{
\item{\code{conn}}{an \linkS4class{H2OConnection}}

\item{\code{job_key}}{a character key representing the identification of the job process.}

\item{\code{model_id}}{the final identifier for the model}
}}
\seealso{
\linkS4class{H2OModel} for the final model types.
}

