% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{H2OModelMetrics-class}
\alias{H2OModelMetrics-class}
\alias{H2OModelMetrics}
\alias{show,H2OModelMetrics-method}
\alias{H2OUnknownMetrics-class}
\alias{H2OBinomialMetrics-class}
\alias{show,H2OBinomialMetrics-method}
\alias{H2OMultinomialMetrics-class}
\alias{show,H2OMultinomialMetrics-method}
\alias{H2OOrdinalMetrics-class}
\alias{show,H2OOrdinalMetrics-method}
\alias{H2ORegressionMetrics-class}
\alias{show,H2ORegressionMetrics-method}
\alias{H2OClusteringMetrics-class}
\alias{show,H2OClusteringMetrics-method}
\alias{H2OAutoEncoderMetrics-class}
\alias{show,H2OAutoEncoderMetrics-method}
\alias{H2ODimReductionMetrics-class}
\alias{show,H2ODimReductionMetrics-method}
\alias{H2OWordEmbeddingMetrics-class}
\alias{H2OCoxPHMetrics-class}
\title{The H2OModelMetrics Object.}
\usage{
\S4method{show}{H2OModelMetrics}(object)

\S4method{show}{H2OBinomialMetrics}(object)

\S4method{show}{H2OMultinomialMetrics}(object)

\S4method{show}{H2OOrdinalMetrics}(object)

\S4method{show}{H2ORegressionMetrics}(object)

\S4method{show}{H2OClusteringMetrics}(object)

\S4method{show}{H2OAutoEncoderMetrics}(object)

\S4method{show}{H2ODimReductionMetrics}(object)
}
\arguments{
\item{object}{An \code{H2OModelMetrics} object}
}
\description{
A class for constructing performance measures of H2O models.
}
