% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{range.H2OFrame}
\alias{range.H2OFrame}
\title{Range of an H2O Column}
\usage{
\method{range}{H2OFrame}(..., na.rm = TRUE)
}
\arguments{
\item{...}{An H2OFrame object.}

\item{na.rm}{ignore missing values}
}
\description{
Range of an H2O Column
}
\examples{
\dontrun{
library(h2o)
h2o.init()

frame <- h2o.createFrame(rows = 6, cols = 2,
                         categorical_fraction = 0.0, 
                         missing_fraction = 0.7, 
                         seed = 123)
range(frame, na.rm = TRUE)
}
}
