% Generated by roxygen2 (4.0.2): do not edit by hand
\name{str.H2OFrame}
\alias{str.H2OFrame}
\title{Describe an H2OFrame object}
\usage{
\method{str}{H2OFrame}(object, cols = FALSE, ...)
}
\arguments{
\item{object}{An H2OFrame object.}

\item{cols}{Logical indicating whether or not to do the str for all columns.}

\item{\dots}{Extra args}
}
\description{
Describe an H2OFrame object
}

