% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{summary,H2OGrid-method}
\alias{summary,H2OGrid-method}
\title{Format grid object in user-friendly way}
\usage{
\S4method{summary}{H2OGrid}(object, show_stack_traces = FALSE)
}
\arguments{
\item{object}{an \code{H2OGrid} object.}

\item{show_stack_traces}{a flag to show stack traces for model failures}
}
\description{
Format grid object in user-friendly way
}

